References to all the datasets included in this package.

This is a RefManageR::BibEntry() object that can be manipulated using the RefManageR::RefManageR package functions or the standard citation functions in R.

Usage

bibliography

Format

An object of class BibEntry (inherits from bibentry) of length 58.