The Bowman, Lehoucq, and Mahoney index of democracy for Central America

Contains the trichotomous measure of regime type from Bowman, Kirk, Fabrice Lehoucq, and James Mahoney. 2005. Measuring Political Democracy: Case Expertise, Data Adequacy, and Central America. Comparative Political Studies 38 (8): 939-970. http://cps.sagepub.com/content/38/8/939. Data available at http://www.blmdemocracy.gatech.edu/.

blm

blm_pmm

Format

An object of class tbl_df (inherits from tbl, data.frame) with 505 rows and 7 columns.

Source

Bowman, Kirk, Fabrice Lehoucq, and James Mahoney. 2005. Measuring Political Democracy: Case Expertise, Data Adequacy, and Central America. Comparative Political Studies 38 (8): 939-970. http://cps.sagepub.com/content/38/8/939. Data available at http://www.blmdemocracy.gatech.edu/.

Variables

year

The calendar year.

blm_country

The country name, as in the original dataset. Use country_name instead. Available only for five Latin American countries (Costa Rica, El Salvador, Guatemala, Honduras and Nicaragua) from 1900 to 2000.

blm

Trichotomous index of democracy. 0 = authoritarian, 0.5 = semidemocratic, 1 = democratic. Available only for five Latin American countries (Costa Rica, El Salvador, Guatemala, Honduras and Nicaragua) from 1900 to 2000.

Standard descriptive variables (generated by this package)

extended_country_name

The name of the country in the Gleditsch-Ward system of states, or the official name of the entity (for non-sovereign entities and states not in the Gleditsch and Ward system of states) or else a common name for disputed cases that do not have an official name (e.g., Western Sahara, Hyderabad). The Gleditsch and Ward scheme sometimes indicates the common name of the country and (in parentheses) the name of an earlier incarnation of the state: thus, they have Germany (Prussia), Russia (Soviet Union), Madagascar (Malagasy), etc. For details, see Gleditsch, Kristian S. & Michael D. Ward. 1999. "Interstate System Membership: A Revised List of the Independent States since 1816." International Interactions 25: 393-413. The list can be found at http://privatewww.essex.ac.uk/~ksg/statelist.html.

GWn

Gleditsch and Ward's numeric country code, from the Gleditsch and Ward list of independent states.

cown

The Correlates of War numeric country code, 2016 version. This differs from Gleditsch and Ward's numeric country code in a few cases. See http://www.correlatesofwar.org/data-sets/state-system-membership for the full list.

in_GW_system

Whether the state is "in system" (that is, is independent and sovereign), according to Gleditsch and Ward, for this particular date. Matches at the end of the year; so, for example South Vietnam 1975 is FALSE because, according to Gleditsch and Ward, the country ended on April 1975 (being absorbed by North Vietnam). It is also TRUE for dates beyond 2012 for countries that did not end by then, depsite the fact that the Gleditsch and Ward list has not been updated since.

Note

The blm_pmm version of this data is taken from from Pemstein, Daniel, Stephen A. Meserve, and James Melton. 2013. "Replication data for: Democratic Compromise: A Latent Variable Analysis of Ten Measures of Regime Type." In: Harvard Dataverse. http://hdl.handle.net/1902.1/PMM. The original dataset is found in blm. The blm_pmm data only goes back to 1946, but is otherwise identical to the blm data. It is included here for completeness.

See also

Other democracy: LIED, PIPE, arat_pmm, bmr, bnr, bollen_pmm, doorenspleet, download_fh_electoral, download_fh, download_reign, download_wgi_voice_and_accountability, fh_pmm, gwf_all, hadenius_pmm, kailitz, magaloni, mainwaring, munck_pmm, pacl, peps, pitf, polity_pmm, polyarchy_dimensions, polyarchy, prc_gasiorowski, svolik_regime, uds_2014, ulfelder, utip, vanhanen, wahman_teorell_hadenius

Other PMM replication data: arat_pmm, bollen_pmm, fh_pmm, hadenius_pmm, mainwaring, munck_pmm, pacl, polity_pmm, prc_gasiorowski

Other regional democracy indexes: mainwaring

Other trichotomous democracy indexes: kailitz, mainwaring, prc_gasiorowski, utip