Generates the extended UDS scores from the latest democracy data in this package

Generates the extended UDS scores from the latest democracy data in this package

generate_extended_uds(verbose = FALSE)

Arguments

verbose

Whether to print diagnostic information about what the function is doing, including information from mirt::mirt about the model being fit. Default is FALSE.

Value

A tibble described in extended_uds. Right now there's no flexibility at all to choose the variables to generate this data, though this might change in the future. See vignette("Replicating_and_extending_the_UD_scores") for more info on how to generate these scores for arbitrary measures.

Examples

if (FALSE) {
extended_uds <- generate_extended_uds()
}