The Geddes Wright and Frantz Autocratic Regimes dataset

The dataset described in Barbara Geddes, Joseph Wright, and Erica Frantz. 2014. "Autocratic Breakdown and Regime Transitions: A New Data Set." Perspectives on Politics 12(2): 313-331. The full data and codebook can be downloaded here http://sites.psu.edu/dictators/. The documentation below is derived from the codebook. Some variables apply to the country-year version, some to the case version and some to the "extended" version (extended back before 1945 with info from the case descriptions).

gwf_all

gwf_all_extended

gwf_autocratic

gwf_autocratic_extended

Format

An object of class tbl_df (inherits from tbl, data.frame) with 7954 rows and 20 columns.

Source

Barbara Geddes, Joseph Wright, and Erica Frantz. 2014. "Autocratic Breakdown and Regime Transitions: A New Data Set." Perspectives on Politics 12(2): 313-331. The full data and codebook can be downloaded here http://sites.psu.edu/dictators/.

Details

gwf_cowcode

Correlates of War (CoW) country code

year

Calendar year. In the "extended_yearly" version of the dataset, these have been calculated back before 1945 from the information available in the casenames.

gwf_country

Country name

gwf_casename

Autocratic regime case name (country name and years); these are the units of observation for duration analysis

gwf_regimetype

Regime type: not independent, foreign-occupied, warlord, democracy, provisional, monarchy, personalist, military, party, party-personalist, party-military, military-personalist, party-military-personalist. Only includes non-democratic regime types.

gwf_duration

Duration of regime up to time t. In the "extended" datasets this always starts at 1.

gwf_fail

Binary indicator for regime failure

gwf_disagree

Binary indicator for country-year observations where GWF coding of democracy differs from CGV (here pacl) coding of democracy

gwf_startdate

Day-Month-Year for the calendar date of the autocratic regime start event. Only in gwf_autocratic datasets.

gwf_enddate

Day-Month-Year for the calendar date of the autocratic regime failure event (31-12-2010 for right-censored autocratic regimes). Only in gwf_autocratic datasets.

gwf_spell

Time-invariant duration of autocratic regime

gwf_fail_subs

Categorical variable marking the subsequent regime type. 1: subsequent regime is democracy 2: subsequent regime is autocratic 3: subsequent regime is warlord, foreign-occupied or ceases to exist 0: no regime failure at duration time t ; and regime still in power December 31, 2010

gwf_fail_type

Categorical variable marking how the autocratic regime ends. 1: regime insiders change rules of regime 2: incumbent loses elections 3: no incumbent runs in competitive election won by opponent 4: popular uprising 5: military coup 6: insurgents, revolutionaries, or combatants fighting a civil war 7: foreign imposition or invasion 8: new autocratic leader selected, changes rules, and remains in power 9: state ceases to exist ends or government fails to control most of the country's territory 0: regime still in power on December 31, 2010

gwf_fail_violent

Categorical variable marking the level of violence during the autocratic regime failure event. 1: no deaths 2: 1-25 deaths 3: 26-1000 deaths 4: >1000 0: regime still in power on December 31, 2010

gwf_party

Binary indicator of party regime type (groups party-based, party-personal, party-military, party-personal-military, oligarchy, and Iran 1979-2010)

gwf_personal

Binary indicator of personalist regime type

gwf_military

Binary indicator of military regime type (groups military, military-personal, indirect military)

gwf_monarchy

Binary indicator of monarchy regime type

Standard descriptive variables (generated by this package)

extended_country_name

The name of the country in the Gleditsch-Ward system of states, or the official name of the entity (for non-sovereign entities and states not in the Gleditsch and Ward system of states) or else a common name for disputed cases that do not have an official name (e.g., Western Sahara, Hyderabad). The Gleditsch and Ward scheme sometimes indicates the common name of the country and (in parentheses) the name of an earlier incarnation of the state: thus, they have Germany (Prussia), Russia (Soviet Union), Madagascar (Malagasy), etc. For details, see Gleditsch, Kristian S. & Michael D. Ward. 1999. "Interstate System Membership: A Revised List of the Independent States since 1816." International Interactions 25: 393-413. The list can be found at http://privatewww.essex.ac.uk/~ksg/statelist.html.

GWn

Gleditsch and Ward's numeric country code, from the Gleditsch and Ward list of independent states.

cown

The Correlates of War numeric country code, 2016 version. This differs from Gleditsch and Ward's numeric country code in a few cases. See http://www.correlatesofwar.org/data-sets/state-system-membership for the full list.

in_GW_system

Whether the state is "in system" (that is, is independent and sovereign), according to Gleditsch and Ward, for this particular date. Matches at the end of the year; so, for example South Vietnam 1975 is FALSE because, according to Gleditsch and Ward, the country ended on April 1975 (being absorbed by North Vietnam). It is also TRUE for dates beyond 2012 for countries that did not end by then, depsite the fact that the Gleditsch and Ward list has not been updated since.

Note

The "extended" datasets *_extended are extended back in time using the gwf_duration variable, so they include years before 1946.The original country-year datasets did not include the period before 1946 explicitly.

See also