Downloads the 2020 update (v. 5.2) of the Lexical Index of Electoral Democracy (annual time series, data to 2019) and processes it using country_year_coder.

Downloads the 2020 update (v. 5.2) of the Lexical Index of Electoral Democracy (annual time series, data to 2019) and processes it using country_year_coder.

redownload_lied(url, verbose = TRUE, return_raw = FALSE, ...)

Arguments

url
The URL of the dataset. Defaults to https://dataverse.harvard.edu/api/access/datafile/:persistentId?persistentId=doi:10.7910/DVN/29106/SXRLK1

verbose
Whether to print a running commentary of what the function is doing while processing the data.

return_raw
Whether to return the raw data, without any processing. Default is FALSE.

...
Other parameters passed to country_year_coder.

Source

Skaaning, Svend-Erik; John Gerring; and Henrikas Bartusevicius (2015). "A Lexical Index of Electoral Democracy." Comparative Political Studies, Vol. 48, No. 12, pp. 1491-1525. Original data and variable descriptions available at https://dataverse.harvard.edu/dataset.xhtml?persistentId=doi:10.7910/DVN/29106

Examples

if (FALSE) {
redownload_lied()}