Milan Svolik's Regime Dataset

The data used to classify democracies and dictatorships for Milan Svolik's The Politics of Authoritarian Rule. The full data and codebook can be downloaded here: https://campuspress.yale.edu/svolik/

svolik_regime

Format

An object of class tbl_df (inherits from tbl, data.frame) with 8977 rows and 11 columns.

Source

M. Svolik. The Politics of Authoritarian Rule. Cambridge: Cambridge University Press, 2012. The full data and codebook can be downloaded here: https://campuspress.yale.edu/svolik/

Variables

svolik_regime_country

The original country name.

svolik_regime_ccode

The original country code (COW).

year

The calendar year.

regime

Regime type: democracy, dictatorship, independence or no authority. From Svolik's codebook:

A dictatorship as an independent country that fails to satisfy at least one of the following two criteria for democracy: (1) free and competitive legislative elections and (2) an executive that is elected either directly in free and competitive presidential elections or indirectly by a legislature in parliamentary systems. By identifying the allocation of executive and legislative powers by competitive elections as the principal difference between dictatorships and democracies, I am following Schumpeter's (1950) and Dahl's (1971) procedural approach to the classification of regime types and build on its operationalization by Alvarez et al. (1996), Boix (2003), and Cheibub et al. (2010).

I nonetheless do depart from Alvarez et al. (1996), Boix (2003), and Cheibub et al. (2010) in two ways. First, I exclude from the data any "no authority" periods (as I just discussed above.) Second, I do not require that the incumbent and the opposition alternate in power before a country is considered democratic. Instead, in order to determine when elections are sufficiently free and competitive for a country to be considered a democracy, I build on the fact that the measurement problem involved in such endeavor systematically differs between transitions from dictatorship to democracy and transitions from democracy to dictatorship.

In the case of transitions from dictatorship to democracy, the measurement problem in ascertaining whether elections are genuinely competitive is most severe when the authoritarian incumbent allows for multi-party elections but gets himself or a candidate he supports elected. The measurement problem is less severe when an opposition candidate wins - in these cases we can be confident that the election was genuinely competitive.

In the case of transitions from democracy to dictatorship, the measurement problem in ascertaining whether elections are genuinely competitive is most severe when an incumbent gets re-elected by significantly manipulating competition in his favor without abolishing multi-party elections. The measurement problem is less severe when an unelected actor (typically the military) replaces or abolishes multi-party elections, the legislature, or the executive - in these cases we can be confident that the country seized to be a democracy.

I therefore employ two sets of coding rules:

The emergence of democracy (transitions from dictatorship to democracy): The country establishes multi-party legislative and multi-candidate executive elections (where applicable) and at least one of the following two conditions is satisfied:

i. A former or current opposition candidate becomes chief executive; or

ii. The authoritarian incumbent (or a candidate supported by the former authoritarian incumbent) gets elected as chief executive but there is a consensus among outside observers that the elections were free and fair.

The breakdown of democracy (transitions from democracy to dictatorship): At least one of the following two conditions is satisfied:

i. An unelected actor replaces or abolishes an election, the legislature, or the executive; or

ii. An incumbent (or a candidate supported by the incumbent) gets elected as chief executive but there is a consensus among outside observers that the elections were not free and fair.

regime_numeric

A numeric version of regime. 1 = democracy, 0 = non-democracy, everything else is NA.

start_year

First year of the regime.

end_year

Last year of the regime.

Standard descriptive variables (generated by this package)

extended_country_name

The name of the country in the Gleditsch-Ward system of states, or the official name of the entity (for non-sovereign entities and states not in the Gleditsch and Ward system of states) or else a common name for disputed cases that do not have an official name (e.g., Western Sahara, Hyderabad). The Gleditsch and Ward scheme sometimes indicates the common name of the country and (in parentheses) the name of an earlier incarnation of the state: thus, they have Germany (Prussia), Russia (Soviet Union), Madagascar (Malagasy), etc. For details, see Gleditsch, Kristian S. & Michael D. Ward. 1999. "Interstate System Membership: A Revised List of the Independent States since 1816." International Interactions 25: 393-413. The list can be found at http://privatewww.essex.ac.uk/~ksg/statelist.html.

GWn

Gleditsch and Ward's numeric country code, from the Gleditsch and Ward list of independent states.

cown

The Correlates of War numeric country code, 2016 version. This differs from Gleditsch and Ward's numeric country code in a few cases. See http://www.correlatesofwar.org/data-sets/state-system-membership for the full list.

in_GW_system

Whether the state is "in system" (that is, is independent and sovereign), according to Gleditsch and Ward, for this particular date. Matches at the end of the year; so, for example South Vietnam 1975 is FALSE because, according to Gleditsch and Ward, the country ended on April 1975 (being absorbed by North Vietnam). It is also TRUE for dates beyond 2012 for countries that did not end by then, depsite the fact that the Gleditsch and Ward list has not been updated since.

See also