The Democracy/Autocracy Dataset by Jay Ulfelder

This is the dataset documented in Jay Ulfelder. Democracy/Autocracy Data Set. 2012. http://hdl.handle.net/1902.1/18836. It comes in two versions: ulfelder, which only goes back to 1955, and ulfelder_extended, which extends the regime classifications back in time using the rgjdura and rgjdurd variables, adding 3861 country-years to the original data, mostly authoritarian regimes.

ulfelder

ulfelder_extended

Format

An object of class tbl_df (inherits from tbl, data.frame) with 7721 rows and 62 columns.

An object of class tbl_df (inherits from tbl, data.frame) with 11582 rows and 62 columns.

Source

Jay Ulfelder. Democracy/Autocracy Data Set. 2012. http://hdl.handle.net/1902.1/18836

Variables with values for all years

pitfcode: PITF Country Code. Three-character country code matching those used by the Political Instability Task Force. Some of these seem off - GER for Germany instead of GMY, UK instead of UKG. I've created ulfelder_scode on the basis of this by converting "GER","MNE","SRB","UK", and "USS" to"GMY","MNT","SER","UKG", and "USR" respectively.

ulfelder_scode: Modified version of pitfcode, matching the Polity IV character codes that appear to be the basis of the PITF codes.

cowcode: Correlates of War (COW) Country Code. Numeric country code matching those used by the Correlates of War project. Missing some. Use cown instead.

year: The year.

rgjtype: Regime Type. Categorical variable indicating the nature of a country's political regime, understood as the institutions regulating access to control over national policy in a sovereign state, at the end of the calendar year.

D = Democracy

A = Non-democracy

NS = Not sovereign. A foreign government or international organization either sets national policy or wields formal veto power

-99 = Country termination (censoring event)

rgjdurd: Consecutive Years of Democracy. Count in years since beginning of an episode of democracy. Coded as missing for countries currently coded as autocracies, interruptions, or interregna, with the exception of the year in which a transition to any of those conditions from democracy occurs.

rgjdura: Time since Democracy. Count in years since last episode of democracy or birth of country, whichever is more recent. Coded as missing for democracies, with the exception of the year in which a transition from autocracy to democracy occurs.

Variables with values only for years greater than 1955

The following variables are all NA for years before 1955 (by construction

I do not impute their values for these years in the ulfelder_extended dataset).

rgjdem: Undocumented.

rgjtrans: Regime Change. Categorical variable indicating the occurrence (or not) of a regime change at some point during the indicated calendar year.

= 1 Transition from non-democracy to democracy

= 0 No change

= -1 Transition from democracy to non-democracy

= -9 Country termination or onset of non-sovereignty

rgjtaut: Transition from Democracy to Autocracy. Dummy variable indicating occurrence of a transition from democracy to autocracy during the calendar year.

= 1 if a transition from democracy to autocracy occurred at any time during the calendar year

= 0 if the country ended the calendar year as a democracy

= -6 if an episode of democracy was ended during the calendar year by the onset of a period of non-sovereignty

= -9 if an episode of democracy ended during the calendar year when the country ceased to exist

= NA if the country was not a democracy at any time during the calendar year

rgjatype: Mode of Transition from Democracy to Autocracy. A categorical variable indicating the nature of the event or process that led in an immediate way to the end of an episode of democracy during that calendar year.

= 1 Military coup: leaders of the state's armed forces claim power

= 2 Self-coup: incumbent officials claim dictatorial powers, cancel or rig elections, or substantially diminish civil liberties

= 3 Rebellion: opposition forces seize power by means other than largely free, fair, and competitive elections

= 4 Other: some other domestic actor—nearly always a monarch—usurps power from an elected government

= -6 Foreign intervention: a foreign government or international organization becomes the ruling authority or a formal veto player

= -9 State termination

= 0 No change (democracy persists)

= NA if the country was not a democracy at any time during the calendar year

rgjaltl: Alternation in Leadership (Democracy Only). Dummy variable indicating whether or not there has been a change in the individual occupying the office of head of government since the start of the current episode of democracy, as of the end of the indicated calendar year. Coded as missing for country-years that ended with any non-democracy.

= 1 No change in chief executive so far

= 0 At least one change in chief executive

rgjaltlt: Alternation in Leadership “Clock” (Democracy Only). Numeric variable counting time until the first change in the head of government.

rgjaltp: Alternation in Party in Power (Democracy Only). Dummy variable indicating whether or not there has been a change in the party holding the office of chief executive since the start of the current episode of democracy, as of the end of the indicated calendar year. In presidential and mixed systems, the president is the reference point. In parliamentary, the premier is tracked. For democracies that start out with coalition governments, any change in the composition of the coalition is treated as a change in the party in power. Coded as missing for country-years that ended with any non-democracy.

= 1 No change in party in power so far

= 0 At least one change in the party in power

rgjaltpt: Alternation in Party in Power “Clock” (Democracy Only). Numeric variable counting time until the first change in party holding the office of chief executive.

rgjtdem: Transition from Autocracy to Democracy. Dummy variable indicating occurrence of a transition from autocracy to democracy during the calendar year

= 1 if a transition from autocracy to democracy occurred at any time during the calendar year

= 0 if the country ended the calendar year as an autocracy

= -6 if an episode of autocracy was ended during the calendar year by the onset of a period of non-sovereignty

= -9 if an episode of democracy ended during the calendar year when the country ceased to exist

= NA if the country was not an autocracy at any time during the calendar year.

rgjepdt: Total Episodes of Democracy. Cumulative count of all episodes of democracy for a given country, including current episode. An episode begins with any onset of democracy (REGDURD = 1); an episode ends with a transition from democracy to autocracy (REGTAUT = 1), an interruption (REGTAUT = -6), an interregnum (REGTAUT = -7), or the end of a country (REGTAUT = -9).

rgjhd: History of Democracy. Dummy variable indicating any occurrence of any episodes of democracy in the country.

= 1 at least one episode of democracy (including current year)

= 0 otherwise

rgjhd2: History of Democracy (Restrictive). Dummy variable indicating that at least one episode of democracy in this country has spanned five or more consecutive calendar years.

= 1 yes (i.e., country has sustained democracy for at least five consecutive years at some point in its history)

= 0 no (i.e., country has never sustained democracy for at least five consecutive years)

rgjhd3: History of Democracy (Broad). Dummy variable indicating at least one episode of democracy in history of current country or parent country.

= 1 if at least one country-year is coded as democracy for the current country or, if current country is the result of a state break-up or partition or was itself sovereign at some previous time, for the "parent" country

= 0 otherwise (i.e., no prior country-years of democracy in current or parent country)

In most instances, this variable will match RGJHD Exceptions will arise for countries that have split, combined, or changed status in the international system in the modern era. For example, East Germany receives a coding of '1' on this variable due to the history of democracy in Germany prior to World War II; Bangladesh is coded '1' on this variable from its inception due to the democratic experience of Pakistan; and Latvia, Lithuania, and Estonia are coded '1' on this variable due to the episodes of democracy that occurred there after they first gained independence after World War I.

rgjyrdt: Total Years of Democracy. Cumulative count of calendar years in which a country was a democracy for at least a portion of the year.

rgjepdl: Longest Previous Episode of Democracy. Three-digit numeric variable indicating length in years of single longest episode of democracy since birth of country. Equal to 0 for countries with no democratic experience or in their first episode of democracy.

rgjepd1: Year of Onset of First Episode of Democracy. Four-digit numeric variable indicating calendar year in which first episode of democracy began (e.g., 1962), including ongoing episode, if it is the first.

rgjepd1a: Age at Onset of First Episode of Democracy. Three-digit numeric variable indicating age of country at time of onset of first democratic episode, including ongoing episode if it is the first. Equal to rgjepd1 minus birth-year of country, so equal to 0 for countries that were "born" with a democratic regime.

rgjtadt: Prior Transitions from Autocracy to Democracy. Cumulative count of all prior transitions from autocracy to democracy in the country in question, including transitions that occurred in the current year.

rgjtdat:Prior Transitions from Democracy to Autocracy. Cumulative count of all transitions from democracy to autocracy in the country in question, including any transitions that occurred in the current year.

rgjtdamc: Undocumented.

rgjtdaac: Undocumented.

rgjeu: Membership in the European Union. Categorical variable indicating membership status in the European Economic Community (EEC), European Community (EC), and European Union (EU), as of 31 December of the listed year. The EEC was formally established in 1957 as the successor to the European Coal and Steel Community, which was formed in 1951. The EU was established in 1992 by the Maastricht Treaty.

= 2 Member

= 1 Formal applicant but not a member

= 0 Neither a member nor a formal applicant

rgjnato: Membership in North Atlantic Treaty Organization (NATO). Categorical variable indicating membership status in NATO, as of 31 December of the listed year. NATO was established in 1949.

= 2 Member

= 1 Formally invited to join but not a member

= 0 Neither a member nor formally invited to join

rgjnatop: Participant in NATO Partnership for Peace. Binary variable indicating participation in NATO’s Partnership for Peace program, as of 31 December of the listed year.

= 1 Participant

= 0 Non-participant

rgjosce: Participant in Conference/Organization for Security and Cooperation in Europe (CSCE/OSCE). Categorical variable indicating participation/membership in CSCE/OSCE, as of 31 December of the listed year. The CSCE was formed in 1973; it became the OSCE in 1995.

= 2 Full member/participant

= 1 Partner for Cooperation

= 0 Non-participant

rgjoecd: Membership in Organization for Economic Cooperation and Development (OECD). Binary variable indicating membership in the OECD, as of 31 December of the listed year. The OECD was founded in 1948 as the Organization for European Economic Cooperation; it was reorganized as the OECD in 1961.

= 1 Member

= 0 Non-member

rgjcoe: Membership in Council of Europe (COE). Binary variable indicating membership in the COE, as of 31 December of the listed year. The Council was founded in 1949.

= 1 Member

= 0 Non-member

rgjcomnw: Membership in the Commonwealth of Nations. Binary variable indicating membership in the COE, as of 31 December of the listed year. Suspended members are still identified as members; expelled members are identified as nonmembers. The Commonwealth was founded in 1931.

= 1 Member

= 0 Non-member

rgjfranc: Member of the International Organization of the French-Speaking World (OIF), or Francophonie. Categorical variable indicating membership status in Francophonie, as of 31 December of the listed year. OIS was formed in 1970.

= 2 Member

= 1 Observer

= 0 Non-member

rgjgenev: Signatory to the Geneva Conventions. Binary variable identifying countries that are signatories to the Geneva Conventions, as of 31 December of the listed year.

= 1 Signatory

= 0 Non-signatory

rgjgattw: Signatory to the General Agreement on Tariffs and Trade (GATT) or Member of the World Trade Organization (WTO). Binary variable identifying countries that were signatories to GATT or are members of the WTO, as of 31 December of the listed year. Countries that were signatories to GATT but did not become members of the WTO in 1995, the year of its establishment, are continuously identified as members by virtue of their continued participation in GATT.

= 1 Signatory/member

= 0 Non-signatory/non-member

rgjapec: Member of the Asia-Pacific Economic Cooperation (APEC). Binary variable indicating membership in APEC, as of 31 December of the listed year. APEC was founded in 1989.

= 1 Member

= 0 Non-member

rgjasean: Member of the Association of Southeast Asian Nations (ASEAN). Binary variable indicating membership in ASEAN, as of 31 December of the listed year. ASEAN was founded in 1967.

= 1 Member

= 0 Non-member

rgjseato: Member of the Southeast Asian Treaty Organization (SEATO). Binary variable indicating membership in SEATO, as of 31 December of the listed year. SEATO was formed in 1954 and was dissolved in 1977.

= 1 Member

= 0 Non-member

rgjoas: Member of the Organization of American States (OAS). Binary variable indicating membership in the OAS, as of 31 December of the listed year. The OAS came into being in 1948, but it originated in 1890 as the International Union of American Republics.

= 1 Member

= 0 Non-member

rgjmerc: Member of the Southern Common Market (Mercado Comun del Sur, or Mercosur). Categorical variable indicating status of membership in Mercosur, as of 31 December of the listed year. Mercosur was founded in 1991 under the Treaty of Asuncion. Observer countries are coded as non-members.

= 2 Member

= 1 Associate member

= 0 Non-member

rgjopec: Member of the Organization of Petroleum Exporting Countries (OPEC). Binary variable indicating membership in OPEC, as of 31 December of the listed year. OPEC was formed in 1960.

= 1 Member

= 0 Non-member

rgjarabl: Member of the Arab League. Binary variable indicating membership in the Arab League, as of 31 December of the listed year. The Arab League was established in 1945.

= 1 Member

= 0 Non-member

rgjoau: Member of the Organization for African Unity (OAU) or African Union. Binary variable indicating membership in the OAU or AU, as of 31 December of the listed year. The OAU was established in 1963; it became the AU in 2001.

= 1 Member

= 0 Non-member

rgjecow: Member of the Economic Community of West African States (ECOWAS). Binary variable indicating membership in ECOWAS, as of 31 December of the listed year. ECOWAS was established in 1975.

= 1 Member

= 0 Non-member

rgjiccpr: Signatory to the International Covenant on Civil and Political Rights. Categorical variable indicating status as signatory to this treaty, as of 31 December of the listed year. The covenant was opened for signature in 1996 and came into force in 1976.

= 2 Ratified, acceded, or succeeded

= 1 Signed only

= 0 No status

rgjiccp1: Signatory to the First Optional Protocol to the ICCPR. Binary variable indicating status as a signatory to the First Optional Protocol, which establishes the right of individuals to petition the Human Rights Committee, as of 31 December of the listed year. The Protocol came into force in 1976.

= 1 Signatory

= 0 Not a signatory

rgjachr: Signatory to the American Convention on Human Rights (a.k.a. the Pact of San Jose, Costa Rica). Categorical variable indicating status as signatory to this treaty, as of 31 December of the listed year. The treaty was adopted in 1969 and entered into force in 1978.

= 2 Ratified, acceded, or succeeded

= 1 Signed only

= 0 No status

rgjachpr: Signatory to the African Charter on Human and People’s Rights. Categorical variable indicating status as signatory to this treaty, as of 31 December of the listed year. The treaty was adopted in 1981 and entered into force in 1986.

= 2 Ratified, acceded, or succeeded

= 1 Signed only

= 0 No status

rgjicj: Declared Recognition as Compulsory the Jurisdiction of the International Court of Justice (ICJ). Binary variable indicating whether the country has declared its recognition of the jurisdiction of the ICJ as compulsory, as of 31 December of the listed year.

= 1 Declared

= 0 No declaration

rgjoic: Member of the Organization of the Islamic Conference (OIC). Binary variable identifying countries that are members of the OIC, which describes itself as "the collective voice of the Muslim world and ensuring to safeguard and protect the interests of the Muslim world in the spirit of promoting international peace and harmony among various people of the world." The OIC was founded in 1969.

= 1 Member state

= 0 Non-member (including observer states)

rgjfdem: Freedom House electoral democracy. Binary variable identifying countries listed by Freedom House as electoral democracies.

= 1 Electoral democracy

= 0 Not an electoral democracy

sftgreg1: Undocumented. Region?

sftgreg2: Undocumented. Region?

sfrgregc: Undocumented. Region?

sftgreg: Undocumented. Region?

Standard descriptive variables (generated by this package)

extended_country_name

The name of the country in the Gleditsch-Ward system of states, or the official name of the entity (for non-sovereign entities and states not in the Gleditsch and Ward system of states) or else a common name for disputed cases that do not have an official name (e.g., Western Sahara, Hyderabad). The Gleditsch and Ward scheme sometimes indicates the common name of the country and (in parentheses) the name of an earlier incarnation of the state: thus, they have Germany (Prussia), Russia (Soviet Union), Madagascar (Malagasy), etc. For details, see Gleditsch, Kristian S. & Michael D. Ward. 1999. "Interstate System Membership: A Revised List of the Independent States since 1816." International Interactions 25: 393-413. The list can be found at http://privatewww.essex.ac.uk/~ksg/statelist.html.

GWn

Gleditsch and Ward's numeric country code, from the Gleditsch and Ward list of independent states.

cown

The Correlates of War numeric country code, 2016 version. This differs from Gleditsch and Ward's numeric country code in a few cases. See http://www.correlatesofwar.org/data-sets/state-system-membership for the full list.

in_GW_system

Whether the state is "in system" (that is, is independent and sovereign), according to Gleditsch and Ward, for this particular date. Matches at the end of the year; so, for example South Vietnam 1975 is FALSE because, according to Gleditsch and Ward, the country ended on April 1975 (being absorbed by North Vietnam). It is also TRUE for dates beyond 2012 for countries that did not end by then, depsite the fact that the Gleditsch and Ward list has not been updated since.

See also