The University of Texas Inequality Project Categorical Dataset of Political Regimes

The University of Texas Inequality Project Categorical Dataset of Political Regimes. Described in S. Hsu. "The Effect of Political Regimes on Inequality, 1963-2002". In: UTIP Working Paper (2008). http://utip.gov.utexas.edu/papers/utip_53.pdf. Data available for download at http://utip.gov.utexas.edu/data/.

Usage

utip

Format

An object of class tbl_df (inherits from tbl, data.frame) with 6215 rows and 18 columns.

Source

S. Hsu. "The Effect of Political Regimes on Inequality, 1963-2002". In: UTIP Working Paper (2008).

Description

This is a categorical dataset of regime types, used in papers on inequality by the UTIP group. The coding rules are not well documented. The category "one party democracy" is particularly difficult to understand. The categories are nevertheless mutually exclusive.

utip_country

The original country name in the dataset.

year

Calendar year

.
other_dictatorship

Whether the regime is a non-military dictatorship. 0 = no, 1 = yes.

military_dictatorship

Whether the regime is a military dictatorship. 0 = no, 1 = yes.

civil_war

Whether the regime is in a civil war. 0 = no, 1 = yes.

one_party_democracy

Whether the regime is a "one party democracy", where "fair" elections are held but only one party competes. 0 = no, 1 = yes.

communist

Whether the regime is communist. 0 = no, 1 = yes.

islamic_republic

Whether the regime is an Islamic republic. 0 = no, 1 = yes. In practice, only Libya and Iran qualify.

social_democracy

Whether the regime is social democracy, using the Esping-Andersen (1990) classification of welfare states. 0 = no, 1 = yes.

conservative_democracy

Whether the regime is a democracy with a conservative or liberal welfare state, using the Esping-Andersen (1990) classification of welfare states. 0 = no, 1 = yes.

utip_dichotomous

Calculated, not in original dataset. Dichotomous index of whether the regime is a social democracy, conservative democracy, or one party democracy. 0 = no, 1 = yes.

utip_dichotomous_strict

Calculated, not in original dataset. Dichotomous index of whether the regime is a social democracy or a conservative democracy (excludes one party democracy). 0 = no, 1 = yes.

utip_trichotomous

Calculated, not in original dataset. Trichotomous index of democracy: 2= regime is a social democracy or a conservative democracy (excludes one party democracy), 1 = regime is a one party democracy, 0 = other.

Standard descriptive variables (generated by this package)

extended_country_name

The name of the country in the Gleditsch-Ward system of states, or the official name of the entity (for non-sovereign entities and states not in the Gleditsch and Ward system of states) or else a common name for disputed cases that do not have an official name (e.g., Western Sahara, Hyderabad). The Gleditsch and Ward scheme sometimes indicates the common name of the country and (in parentheses) the name of an earlier incarnation of the state: thus, they have Germany (Prussia), Russia (Soviet Union), Madagascar (Malagasy), etc. For details, see Gleditsch, Kristian S. & Michael D. Ward. 1999. "Interstate System Membership: A Revised List of the Independent States since 1816." International Interactions 25: 393-413. The list can be found at http://privatewww.essex.ac.uk/~ksg/statelist.html.

GWn

Gleditsch and Ward's numeric country code, from the Gleditsch and Ward list of independent states.

cown

The Correlates of War numeric country code, 2016 version. This differs from Gleditsch and Ward's numeric country code in a few cases. See http://www.correlatesofwar.org/data-sets/state-system-membership for the full list.

in_GW_system

Whether the state is "in system" (that is, is independent and sovereign), according to Gleditsch and Ward, for this particular date. Matches at the end of the year; so, for example South Vietnam 1975 is FALSE because, according to Gleditsch and Ward, the country ended on April 1975 (being absorbed by North Vietnam). It is also TRUE for dates beyond 2012 for countries that did not end by then, depsite the fact that the Gleditsch and Ward list has not been updated since.

See also