Authoritarian Regimes Data Set, version 5.0, by Axel Hadenius, Jan Teorell, & Michael Wahman

The Authoritarian Regimes Data set by Axel Hadenius, Jan Teorell, & Michael Wahman, described in Hadenius, Axel & Jan Teorell. 2007. "Pathways from Authoritarianism", Journal of Democracy 18(1): 143-156 and Wahman, Michael, Jan Teorell, and Axel Hadenius. 2013. Authoritarian regime types revisited: updated data in comparative perspective. Contemporary Politics 19 (1): 19-34. The dataset and codebook can be downloaded from https://sites.google.com/site/authoritarianregimedataset/data. I simply copy the codebook below.

wahman_teorell_hadenius

Format

An object of class tbl_df (inherits from tbl, data.frame) with 8231 rows and 54 columns.

Source

Hadenius, Axel and Jan Teorell. 2005. "Assessing Alternative Indices of Democracy", C&M Working Papers 6, IPSA, August 2005.

Hadenius, Axel and Jan Teorell. 2006. "Authoritarian Regimes: Stability, Change, and Pathways to Democracy, 1972-2003", University of Notre Dame, Kellogg Institute Working Paper Series 331, November 2006 (http://kellogg.nd.edu/publications/workingpapers/WPS/331.pdf).

Wahman, Michael, Jan Teorell, and Axel Hadenius. 2013. Authoritarian regime types revisited: updated data in comparative perspective. Contemporary Politics 19 (1): 19-34

Hadenius, Axel and Jan Teorell. 2007. "Pathways from Authoritarianism", Journal of Democracy 18(1): 143-156.

NOTE ON VERSION 4

In this update between version 2.0 and 3.0 (October 10, 2007) version of the data, we have corrected some erroneous codings in the regime type variables for Czechoslovakia (1992), Djibouti (1992-2005), Kyrgyz Republic (1991-1994), Lesotho (1993-1997), Singapore (1972-1980), Tajikistan (1991-1994 and 2005), and Uzbekistan (1991-1993). We also updated the Polity 4 data for the year of 2005 (instead of using imputed scores), which affected the regime classification of the Philippines for that year. Between version 3.0 and 4.0 codings were adjusted for Fiji in the period 2007-2008.

For each variable in the database n refers to the number of observed country years, N to the number of countries covered.

Country and time coverage

This data set covers the time period 1972-2010 and includes all 192 nations recognized as members of the UN except the four "micro states" of Europe (Andorra, Liechtenstein, Monaco and San Marino) and two "micro states" in the Pacific that are not members of the World Bank (Nauru and Tuvalu). In addition, we have included Taiwan and seven states that have dissolved or merged with other states: the USSR, Czechoslovakia, the Federal People's Republic of Yugoslavia, East Germany, North and South Yemen, and South Vietnam. This makes a tally of 195 states. We have treated Germany as a continuation of West Germany, Vietnam as a continuation of North Vietnam, and Ethiopia as a continuation of itself before the secession of Eritrea. By contrast, we treat the Republic of Yemen, the Federal Republic of Yugoslavia (Serbia and Montenegro) and the Russian Federation as new states. In the case of Yugoslavia (Serbia and Montenegro) Serbia is seen as the continuation of this state following the independence of Montenegro. Cyprus refers to the Greek part of Cyprus. Albania is treated as the same case after the independence of Kosovo. Kosovo is treated as a new case from 2009.

Identification variables

wahman_teorell_hadenius_country

(n: 7991, N: 195) Country name. Lists the full country name.

siffra

(n: 7991, N: 195) Country code. Numeric country code.

wahman_teorell_hadenius_cow

(n:7150, N: 192) Corralates of war country code. Source: Corralates of war 2 project http://www.correlatesofwar.org/.

ccode

(n:5867, N: 168) Polity country code (polity IV) Example: 85- Kenya 200- United Kingdom Changes made in the country codes when incorporating them into our database: Germany West (original code: 260) = Germany (ccode 255) Ethiopia (original code: 530) = Ethiopia (ccode 529) North Vietnam (original code: 816) = Vietnam (ccode 818) Russia (original code: 365) before 1992 = USSR (ccode 364) Source: Marshall, Monty G & Jaggers, Keith: Polity IV project, Integrated Network for societal Conflict research (INSCR) Program Center for International Development and Conflict Management (CIDCM) University of Maryland, Collage Park 20742, http://www.cidcm.umd.edu/inscr/polity.

ccode_qog

(n:7527, N:193) ISO country code. In the following cases the ISO country numbers has been changed to match the countrycoding principles of our database: Germany West before 1991 (original code: 280) = Germany (ccode 276) Ethiopia before 1993 (original code: 230) = Ethiopia (ccode 231) North Vietnam before 1976 (original code: 998) = Vietnam (ccode 704) Source: http://www.iso.org/en/prods-services/iso3166ma/02iso-3166-code-lists/index.html, http://en.wikipedia.org/wiki/ISO3166-1

year

(n: 7991, N: 195) Year of observation.

Indices of democracy

fhadd

(n: 6712, N: 195) Average Freedom House political rights and civil liberty scores scaled 0-10, where 10 indicate the highest degree of political rights and civil liberties and 0 the lowest degree. Source: The Freedom House surveys (Freedom in the world). www.freedomhouse.org.

revpol2

(n: 5794, N:170) Revised combined polity score (polity2), scaled 0-10, where 10 indicates strongly democratic and 0 strongly autocratic. Source: Marshall, Monty G & Jaggers, Keith: Polity IV project, Integrated Network for societal Conflict research (INSCR) Program Center for International Development and Conflict Management (CIDCM) University of Maryland, Collage Park 20742, http://www.systemicpeace.org/polity/polity4.htm.

fhpol

(n: 5789, N: 170) Average Polity (revpol2) & Freedom House (fhadd) scores, scaled 0-10.1 Source: See fhadd and revpol2.

ifhpol

(n: 6712, N: 195) Imputed average Polity (revpol2) & Freedom House (fhadd) scores (scaled 0-10), where missing values have been imputed by regressing the fhpol index on the Freedom House scores (fhadd), which have better country coverage than Polity.2 Countries with an ifhpol score larger than 7.0 are coded as democracies. 1 As we have shown in a recent paper, this combined FH/Polity index outperforms all rival indices of democracy in an independent assessment (Hadenius and Teorell 2005). 2 Imputed scores larger than 10 are set equal to 10. This imputation in effect also affects the time-series information on ifhpol for countries with some Polity data (although with less coverage than FH). In rare instances, this means that ifhpol in these countries changes although the underlying FH scores do not change. Source: See fhadd and revpol2.

Regime classification variables

regime1ny

(n: 6693, N:195) Collapsed regime type 1 Monarchy 2 Military 3 One party 4 Multi-party 9 No-party 99 Other 100 Democracy Using the mean of the Freedom House and Polity scales (ifhpol), we draw the line between democracies and autocracies at 7.0. We chose his threshold value by estimating the mean cutoff point separating democracy from autocracy in five well-known categorical measures of democracy: Cheibub et al. (2010), Boix et al (Forthcoming) and Bernhard et al. (2001)3, together with Polity's own categorical threshold for "democracy" and Freedom House's threshold for "Electoral Democracy." At the core of our typology of authoritarian regime types is a distinction between three different modes of political power maintenance (probably the three most widely used throughout history): 1. Hereditary succession, or lineage, corresponding to Monarchies; we define monarchies as those regimes in which a person of royal descent has inherited the position of head of state in accordance with accepted practice and/or the constitution (one cannot proclaim oneself a monarch). It bears stressing that we only apply this classification to countries where the sovereign exercises real political power; ceremonial monarchies are thus excluded. This occurs in Afghanistan 1979 and 2001, Bosnia-Herzegovina 1995, Cambodia 1979, Lebanon 1990, and Solomon Islands 2003. Users particularly interested in the time-series information on democracy for these countries should consider more case-specific imputation techniques. 2. The actual or threatened use of military force, referring to Military regimes, where the armed forces may exercise political power either directly or indirectly (i.e., by controlling civilian leaders behind the scenes). Regimes where persons of military background are chosen in open elections (which have not been controlled by the military) thus should not count as military. "Rebel regimes" form a special subcategory. They include cases where a rebel movement (one not formed out of the regular armed forces) has taken power by military means, and the regime has not as yet been reconstituted as another kind of regime. 3. Popular elections, designating the various electoral regimes. We distinguish among three basic types of electoral regimes. The first is the No-Party Regime, where elections are held but all political parties (or at least any candidate representing a party) are prohibited. Elections in no-party regimes may display an element of competition, but thus only among individual candidates. Second, in One-Party Regimes, all parties but one is forbidden (formally or de facto) from taking part in elections. A small number of non-party candidates may also be allowed to take part and get elected; there may be satellite parties which are autonomous in name, but which cannot take an independent position; and competition between candidates from the same (ruling) party may also obtain; we still code the regime one-party. It is not enough, moreover, that a regime calls itself a one-party state; elections in such a structure must also be held. Third, and finally, we define Limited Multiparty regimes as regimes that hold parliamentary or presidential elections in which (at least some) candidates are able to participate who are independent of the ruling regime. This classification holds even when opposition parties refrain voluntarily from taking part in elections. It also embraces cases where parties are absent, but where this is not the result of any prohibition against party activities: the candidates in question have simply chosen to stand for election as individuals. These latter we classify as Party-Less limited multiparty systems. Finally, we have a residual category called others, including a few cases that do not fit under any other regime type, given the definitions applied. The categories in regime1ny are not mutually exclusive. All monarchical regimes with amalgams (regimeny=16, 17, 23 or 24) are treated as monarchies, all military regimes with sub-types and amalgams (regimeny=4, 5, 6, 7 or 18) are treated as military regimes, and multiparty regimes with sub-types are treated as multiparty regimes (regimeny =1 or 2). Only pure no-party (regimeny=3) and one-party (regimeny=8) regimes are treated as no-party and one-party regimes, respectively. The minor types (regimeny=9, 19, 20, 21, 22) are treated as other. Our regime classification pertains to December 31 as of each year. Sources: Banks and Wilson (2012), supplemented with: Elections in the world http://www.electionworld.org, IFES Election guide http://209.50.195.230, Journal of democracy: Election watch http://muse.jhu.edu/journals/journal_of_democracy/election_watch/, Parties and elections in Europe http://www.parties-and-elections.de/index.html, Freedom House "Freedom in the World Reports" http://www.freedomhouse.org, The interparliamentary Union http://www.ipu.org/english/home.htm, Political Database of the Americas http://www.georgetown.edu/pdba/, Rulers http://www.rulers.org, Keesing's Record of World Events http://keesings.gvpi.net/keesings/lpext.dll?f=templates&fn=main-h.htm, The Swedish Institute of International affairs, Country guides http://www.landguiden.se, The US Library of congress Country Studies (Federal Research Division of Library of Congress) http://memory.loc.gov/frd/cs/cshome.html#toc, Banks and Mueller: Political handbook of the world 1979, The Europe world tear book, various years, The Economist Intelligence Unit: Quaterly Economic Reviews, Country profiles and Country Reports, various years, The CIA World Fact book http://www.cia.gov/cia/publications/factbook, Country Reports on Human Rights practices (Reports submitted to Congress by US Department of State), various years, later editions on http://www.state.gov/g/drl/hr/c1470.htm, Encycopledia Britannica online http://search.eb.com.

regime1nyrobust

Same as regime1ny, but with 7.5 democracy threshold.

regimeny

(n: 6693, N:195) Regime type 1 Limited Multiparty 2 Partyless 3 No-Party 4 Military 5 Military No-Party 6 Military Multiparty 7 Military One-party 8 One-Party 9 Other 16 One-Party Monarchy 17 Monarchy 18 Rebel Regime 19 Civil War 20 Occupation 21 Theocracy 22 Transitional Regime 23 No-Party Monarchy 24 Multiparty Monarchy 100 Democracy Based on the classifications in regime1ny we also code hybrids (or amalgams) combining elements from more than one regime type. Monarchies may carry out elections in various forms: multiparty elections, no-party elections, and also one-party elections. The same goes for military regimes. In addition to the main types and their amalgams, we have identified several minor types of authoritarian regime. In a theocracy, decisive political power lies in the hands of a religious elite. Temporary regimes, the purpose of which is to carry out a transition, are classified as transitional regimes.7 There are furthermore countries in which the official government does not in reality control the territory. This may be due to civil war or occupation by foreign troops. Source: See regime1ny

regimenyrobust

Same as regimeny, but with 7.5 democracy threshold.

mon

(n: 6693, N:195) Monarchy main type dummy 0- Not a monarchy 1- Monarchy Coded as 1 for all monarchies, regardless of whether they are amalgams or sub-types (regimeny=16, 17, 23 or 24), 0 otherwise. Source: See regime1ny.

monrobust

Same as mon, but with 7.5 democracy threshold.

mil

(n: 6693, N:195) Military main type dummy 0- civilian rule 1- military rule Coded as 1 for all military regimes, regardless of whether they are amalgams or sub-types (regimeny=4, 5, 6, 7 or 18), 0 otherwise. Source: See regime1ny.

milrobust

Same as mil, but with 7.5 democracy threshold.

mul

(n: 6693, N:195) multiparty main type dummy 0- Not multiparty 7 A transitional regime can only last in our schema for up to three years; after that, it is given a different and more fitting classification. 1- Limited Multiparty Coded as 1 for all limited multiparty regimes, regardless of whether they are amalgams or sub-types (regimeny=1, 2, 6, 24 or 25), 0 otherwise. Source: See regime1ny.

mulrobust

Same as mul, but with 7.5 democracy threshold.

onep

(n: 6693, N:195) one-party main type dummy 0- Not a one party regime 1- One-party regime Coded as 1 for all one-party regimes, regardless of whether they are amalgams or sub-types (regimeny=7, 8 or 16), 0 otherwise. Source: See regime1ny.

oneprobust

Same as onep, but with 7.5 democracy threshold.

nop

(n: 6693, N:195) noparty main type dummy 0- Not a no-party system 1- No-party system Coded as 1 for all no-party regimes, regardless of whether they are amalgams or sub-types (regimeny=3, 5 or 23), 0 otherwise. Source: See regime1ny.

noprobust

Same as nop, but with 7.5 democracy threshold.

partsz

(n:6504, N:194) size of the largest party in legislature (in fractions) Counts the largest parties' number of seats divided with the legislative assembles' total number of seats expressed in fractions. In countries with a two-chamber parliament the lower house is counted. Source: Banks and Wilson (2012), Elections in the world http://www.electionworld.org, Beck et al (2001), IFES Election guide http:// 209.50.195.230, Journal of democracy: Election watch http://muse.jhu.edu/journals/journal_of_democracy/election_watch/, Parties and elections in Europe http://www.parties-and-elections.de/index.html, The interparliamentary Union http://www.ipu.org/english/home.htm, Political Database of the Americas http://www.georgetown.edu/pdba/, Rulers http://www.rulers.org, Keesing's Record of World Events http://keesings.gvpi.net/keesings/lpext.dll?f=templates&fn=main-h.htm, The CIA World Fact book http://www.cia.gov/cia/publications/factbook

partsz1

(n:6504, N:193) size of largest party (in fraction), zero for one party regime Codes all one-party regimes as 0 instead of 1 as is done in partsz, otherwise this variable corresponds to the former variable partsz. When the degree of "dominance" of the largest party within multiparty regimes is to be controlled for, this variable should be used. Source: See partsz.

Historical variables and regime duration indicators

yrindep

(n:7605, N:195) Year of independence (1700=1700 or before) We have generally used the data when a state became independent or unified within the (more or less) current borders. This pertains for instance to Hungary, which is coded as independent in 1918 and not in 1001 when it was first unified. Cases of occupation when the occupant power never established any civilian rule has not been counted as cases of colonisation. All countries that became independent in 1700 or prior to this year is coded as yrindep=1700. Source: CIA World Fact Book and Encyclopedia Britannica Online.

yrterm

(n:7605, N:195) Year of termination of the country (2005= still existing at the end of our observation period) The following cases has a noted year of termination<2005: East Germany (1990), Czechoslovakia (1993), USSR (1991), South Yemen (1990), North Yemen (1990), Yugoslavia FPR (1992), South Vietnam (1976). Source: CIA World Fact Book and Encyclopedia Britannica Online.

regstart

(n: 2673, N: 194) Starting year of regime (backdated to 1960) Indicates the starting year of the current regime type, backdated until 1960 for the regime that was in place in 1972, that is, at the start of our observation period. The variable is only coded for the first regime period of each country.

lagregime1ny

(n: 6693, N:194) Previous regime type (collapsed regime type) 1 Monarchy 2 Military 3 One party 4 Multi-party 9 No-party 99 Other 100 Democracy Indicates the current regimes' previous regime type according to the classification of regime1ny. At the first year of the time-series each country has its current regime type as lagregime1ny. For definitions of regime types, see regime1ny.. Source: See regime1ny.

lagregime1nyrobust

Same as lagregime1ny, but with 7.5 democracy threshold.

lagregimeny

(n: 6693, N:194) Previous regime type 1 Limited Multiparty 2 Partyless 3 No-Party 4 Military 5 Military No-Party 6 Military Multiparty 7 Military One-party 8 One-Party 9 Other 16 One-Party Monarchy 17 Monarchy 18 Rebel Regime 19 Civil War 20 Occupation 21 Theocracy 22 Transitional Regime 23 No-Party Monarchy 24 Multiparty Monarchy 25 Multiparty Occupied 100 Democracy Indicates the current regimes' previous regime type according to the classification of regimeny. At the first year of the time-series each country has its current regime type as lagregimeny. For definitions of regime types, see regimeny. Source: See regime1ny.

lagregimenyrobust

Same as lagregimeny, but with 7.5 democracy threshold.

regnumb1ny

(n: 6693, N:195) Number of consecutive regime periods (collapsed regime type) Indicates the consecutive number of the current regime period for a country according to the classification of regime1ny. Source: See regime1ny

regnumb1nyrobust

Same as regnumb1ny, but with 7.5 democracy threshold.

regnumbny2

(n: 6693, N:195) Number of consecutive regime periods Indicates the consecutive number of the current regime period for a country according to the classification of regimeny. Source: See regime1ny

regnumbny2robust

Same as regnumbny2, but with 7.5 democracy threshold.

totdur1ny

(n: 6693, N:195) Duration of current regime period (collapsed regime type) Indicates the number of years that the current regime, according to the classification of regime1ny, has been in place in a country (up until 2003, the end of the observation period). Regime durations are backdated to 1960 according to regstart. Source: See regime1ny

totdur1nyrobust

Same as totdur1ny, but with 7.5 democracy threshold.

totdurny2

(n: 6693, N:195) Duration of current regime period (in years) Indicates the number of years that the current regime, according to the classification of regimeny, has been in place in a country (up until 2003, the end of the observation period). Regime durations are backdated to 1960 according to regstart. Source: See regime1ny

totdurny2robust

Same as regnumbny2, but with 7.5 democracy threshold.

Indicators of personalism

persagg1ny

(n: 6693, N:195) Mean executive turnover (collapsed regime type) Measures the total number of changes of the chief executive during the regime spell divided by the years of regime spell duration, according to the classification of regime1ny. The effective executive may be the president, prime minister, leader of the ruling party, the monarch or the ruling military junta, or someone else, working behind political figure heads. Source: Banks and Wilson (2012), Elections in the world http://www.electionworld.org, Rulers http://www.rulers.org, Keesing's Record of World Events http://keesings.gvpi.net/keesings/lpext.dll?f=templates&fn=main-h.htm, The Swedish Institute of International affairs: Country guides http://www.landguiden.se, Country Reports on Human Rights practices (Reports submitted to Congress by US Department of State), Encyclopedia Britannica http://search.eb.com

persagg1nyrobust

Same as persagg1ny, but with 7.5 democracy threshold.

persaggny2

(n: 6693, N:195) Mean executive turnover (within regime spells) Same as persagg1ny, but according to the classification of regimeny.

persaggny2robust

Same as persaggny2, but with 7.5 democracy threshold. Source: See persagg1ny.

tenure1ny

(n: 6693, N:195) Mean years of executive tenure (collapsed regime type) Measures the years of regime spell duration divided by the total number of changes of the executive during the regime spell, according to the classification of regime1ny. In case no change of executive occurred during a regime spell, tenure1ny is set equal to the regime spell duration. Source: See persaggny2.

tenure1nyrobust

Same as tenure1ny, but with 7.5 democracy threshold.

tenureny2

(n: 6693, N:195) Mean years of executive tenure (within regime spells) Same as tenure1ny, but according to the classification of regimeny. Source: See persagg1ny.

tenureny2robust

Same as tenureny2, but with 7.5 democracy threshold.

Standard descriptive variables (generated by this package)

extended_country_name

The name of the country in the Gleditsch-Ward system of states, or the official name of the entity (for non-sovereign entities and states not in the Gleditsch and Ward system of states) or else a common name for disputed cases that do not have an official name (e.g., Western Sahara, Hyderabad). The Gleditsch and Ward scheme sometimes indicates the common name of the country and (in parentheses) the name of an earlier incarnation of the state: thus, they have Germany (Prussia), Russia (Soviet Union), Madagascar (Malagasy), etc. For details, see Gleditsch, Kristian S. & Michael D. Ward. 1999. "Interstate System Membership: A Revised List of the Independent States since 1816." International Interactions 25: 393-413. The list can be found at http://privatewww.essex.ac.uk/~ksg/statelist.html.

GWn

Gleditsch and Ward's numeric country code, from the Gleditsch and Ward list of independent states.

cown

The Correlates of War numeric country code, 2016 version. This differs from Gleditsch and Ward's numeric country code in a few cases. See http://www.correlatesofwar.org/data-sets/state-system-membership for the full list.

in_GW_system

Whether the state is "in system" (that is, is independent and sovereign), according to Gleditsch and Ward, for this particular date. Matches at the end of the year; so, for example South Vietnam 1975 is FALSE because, according to Gleditsch and Ward, the country ended on April 1975 (being absorbed by North Vietnam). It is also TRUE for dates beyond 2012 for countries that did not end by then, depsite the fact that the Gleditsch and Ward list has not been updated since.

See also