The Lexical Index of Democracy, v. 5.2

The Lexical Index of Democracy described in Skaaning, Svend-Erik; John Gerring; and Henrikas Bartusevicius (2015). "A Lexical Index of Electoral Democracy." Comparative Political Studies, Vol. 48, No. 12, pp. 1491-1525. Original data and variable descriptions available at https://dataverse.harvard.edu/dataset.xhtml?persistentId=doi:10.7910/DVN/29106.

LIED

Format

An object of class tbl_df (inherits from tbl, data.frame) with 21365 rows and 19 columns.

Source

Skaaning, Svend-Erik; John Gerring; and Henrikas Bartusevicius (2015). "A Lexical Index of Electoral Democracy." Comparative Political Studies, Vol. 48, No. 12, pp. 1491-1525. Original data and variable descriptions available at https://dataverse.harvard.edu/dataset.xhtml?persistentId=doi:10.7910/DVN/29106

Description of the Construction of LIED

To code the lexical index we make use of five variables developed initially in the Political Institutions and Events (PIPE) dataset (Przeworski et al. 2013): LEGSELEC, EXSELEC, OPPOSITION, MALE SUFFRAGE, and FEMALE SUFFRAGE. Since PIPE does not attempt to measure the quality of elections, we generate a sixth variable: COMPETITION. All variables are binary, coded 1 if the following circumstances obtain, and 0 otherwise.

lied_country

The country name used in the original dataset. 2 country-years were repeated in the original data and have been dropped.

lied_cow

The COW country code used in the original dataset.

vdem

The V-Dem country code.

year

Calendar year.

legselec

A legislative body issues at least some laws and does not perform executive functions. The lower house (or unicameral chamber) of the legislature is at least partly elected. The legislature has not been closed.

exselec

The chief executive is either directly or indirectly elected (i.e., chosen by people who have been elected)

opposition

The lower house (or unicameral chamber) of the legislature is (at least in part) elected by voters facing more than one choice. Specifically, parties are not banned and (a) more than one party is allowed to compete or (b) elections are nonpartisan (i.e., all candidates run without party labels)

male_suffrage

Virtually all male citizens are allowed to vote in national elections. Legal restrictions pertaining to age, criminal conviction, incompetence, and local residency are not considered. Informal restrictions such as those obtaining in the American South prior to 1965 are also not considered.

female_suffrage

Virtually all female citizens are allowed to vote in national elections. Similar coding rules apply.

competition

The chief executive offices and seats in the effective legislative body are filled by elections characterized by uncertainty (see Przeworski 2000: 16- 17), meaning that the elections are, in principle, sufficiently free to enable the opposition to gain power if they were to attract sufficient support from the electorate. This presumes that control over key executive and legislative offices is determined by elections, the executive and members of the legislature have not been unconstitutionally removed, and the legislature has not been dissolved. With respect to the electoral process, this presumes that the constitutional timing of elections has not been violated (in a more than marginal fashion), non-extremist parties are not banned, opposition candidates are generally free to participate, voters experience little systematic coercion in exercising their electoral choice, and electoral fraud does not determine who wins. With respect to the outcome, this presumes that the declared winner of executive and legislative elections reflects the votes cast by the electorate, as near as can be determined from extant sources. Incumbent turnover (as a result of multi-party elections) is regarded as a strong indicator of competition, but is neither necessary nor sufficient. In addition, we rely on reports from outside observers (as reported in books, articles, and country reports) about whether the foregoing conditions have been met in a given election (see Svolik 2012: 24). Coding for this variable does not take into account whether there is a level playing field, whether all contestants gain access to funding and media, whether media coverage is unbiased, whether civil liberties are respected, or other features associated with fully free and fair elections. COMPETITION thus sets a modest threshold.

lexical_index

To generate the lexical index from these six binary variables, a country-year is assigned the highest score (L0-6) for which it fulfills all requisite criteria, as follows:L0: LEGSELEC=0 & EXSELEC=0.L1: LEGSELEC=1 or EXSELEC=1.L2: LEGSELEC=1 & OPPOSITION=1.L3: LEGSELEC=1 & OPPOSITION=1 & EXSELEC=1.L4: LEGSELEC=1 & OPPOSITION=1 & EXSELEC=1 & COMPETITION=1.L5: LEGSELEC=1 & OPPOSITION=1 & EXSELEC=1 & COMPETITION=1 & (MALE SUFFRAGE=1 or FEMALE SUFFRAGE=1).L6: LEGSELEC=1 & OPPOSITION=1 & EXSELEC=1 & COMPETITION=1 & MALE SUFFRAGE=1 & FEMALE SUFFRAGE=1.In a small number of cases, the resulting index appears to have been incorrectly constructed:
## # A tibble: 28 x 4
##    lied_country  year lexical_index lexical_index_original
##    <chr>        <dbl>         <dbl>                  <dbl>
##  1 Netherlands   1917             5                      4
##  2 San Marino    1921             3                      1
##  3 San Marino    1922             3                      1
##  4 Australia     1876             2                      4
##  5 Australia     1877             2                      4
##  6 Australia     1878             2                      4
##  7 Australia     1879             2                      4
##  8 Australia     1880             2                      4
##  9 Australia     1881             2                      4
## 10 Australia     1882             2                      4
## # ... with 18 more rows

I've recalculated the index for these cases and put the original value in the lexical_index_original variable.

Standard descriptive variables (generated by this package)

extended_country_name

The name of the country in the Gleditsch-Ward system of states, or the official name of the entity (for non-sovereign entities and states not in the Gleditsch and Ward system of states) or else a common name for disputed cases that do not have an official name (e.g., Western Sahara, Hyderabad). The Gleditsch and Ward scheme sometimes indicates the common name of the country and (in parentheses) the name of an earlier incarnation of the state: thus, they have Germany (Prussia), Russia (Soviet Union), Madagascar (Malagasy), etc. For details, see Gleditsch, Kristian S. & Michael D. Ward. 1999. "Interstate System Membership: A Revised List of the Independent States since 1816." International Interactions 25: 393-413. The list can be found at http://privatewww.essex.ac.uk/~ksg/statelist.html.

GWn

Gleditsch and Ward's numeric country code, from the Gleditsch and Ward list of independent states.

cown

The Correlates of War numeric country code, 2016 version. This differs from Gleditsch and Ward's numeric country code in a few cases. See http://www.correlatesofwar.org/data-sets/state-system-membership for the full list.

in_GW_system

Whether the state is "in system" (that is, is independent and sovereign), according to Gleditsch and Ward, for this particular date. Matches at the end of the year; so, for example South Vietnam 1975 is FALSE because, according to Gleditsch and Ward, the country ended on April 1975 (being absorbed by North Vietnam). It is also TRUE for dates beyond 2012 for countries that did not end by then, depsite the fact that the Gleditsch and Ward list has not been updated since.

Deviations from PIPE

The LIED compilers state that:

Although we employ PIPE as an initial source for coding LEGSELEC, EXSELEC, OPPOSITION, MALE SUFFRAGE, and FEMALE SUFFRAGE, we deviate from PIPE—based on our reading of countryspecific sources—in several ways. First, with respect to executive elections, in the PIPE dataset "Prime ministers are always coded as elected if the legislature is open." However, for our purposes we need an indicator that also takes into account whether the government is responsible to an elected parliament if the executive is not directly elected—a situation generated by a number of European monarchies prior to World War I, by episodes of international supervision such as Bosnia-Herzegovina in the first years following the civil war, and by some monarchies in the Middle East and elsewhere (e.g., Liechtenstein, Monaco, and Tonga) in the contemporary era. To illustrate, PIPE codes Denmark as having executive elections from 1849 to 1900 although the parliamentary principle was not established until 1901. Before then, the government was accountable to the king. Among the current cases with elected multiparty legislatures not fulfilling this condition, we find Jordan and Morocco. In order to achieve a higher level of concept-measure consistency, we have thus recoded all country-years (based on country-specific accounts) for this variable where our sources suggested doing so.

We also conduct original coding for countries whose coding is incomplete in PIPE and for additional countries such as the German principalities that are not covered in PIPE. In this fashion, we generate a complete dataset for all six variables covering all independent countries of the world in the period under study (1800-2013 or 1789-2019 for v. 5.2). Whereas the numbers of observations for the PIPE variables range between 14,465 and 15,302, our dataset provides 18,142 observations for all variables. Except for minor adjustments regarding executive elections (mentioned above), this additional coding follows the rules laid out in the PIPE codebook. Coding decisions are based on country-specific sources that are too numerous to specify. In rare instances we stumbled upon information that required a re-coding of PIPE variables, so the two datasets do not correspond exactly.

Countries are coded across these conditions for the length of their sovereign existence within the 1800-2013 (1789-2019 for v. 5.2) timespan, generating a dataset with 221 countries. To identify independent countries we rely on Gleditsch (2013) and Correlates of War (2011), supplemented from 1800 to 1815 by various country-specific sources. Importantly, electoral democracy does not presume complete sovereignty. A polity may be constrained in its actions by other states, by imperial control (as over a colony), by international treaties, or by world markets. Thus, to say that a polity is an electoral democracy is to say that it functions as such for policies over which it enjoys decisionmaking power. Scores for each indicator reflect the status of a country on the last day of the calendar year (31 December) and are not intended to reflect the mean value of an indicator across the previous 364 days.

See also