The Anckar-Fredriksson dataset of political regimes

Dataset described in C. Anckar and C. Fredriksson. "Classifying political regimes 1800-2016: a typology and a new dataset". In: European Political Science (2018). DOI: 10.1057/s41304-018-0149-8. https://doi.org/10.1057/s41304-018-0149-8. Data and codebook at the link. This is version 2 of the data, available at https://dataverse.harvard.edu/dataset.xhtml?persistentId=doi:10.7910/DVN/7SSSAH&version=2.0.

anckar

Format

An object of class tbl_df (inherits from tbl, data.frame) with 18777 rows and 13 columns.

Source

C. Anckar and C. Fredriksson. "Classifying political regimes 1800-2016: a typology and a new dataset". In: European Political Science (2018). DOI: 10.1057/s41304-018-0149-8. https://doi.org/10.1057/s41304-018-0149-8. Version 2 available at https://dataverse.harvard.edu/dataset.xhtml?persistentId=doi:10.7910/DVN/7SSSAH&version=2.0.

Variables

anckar_country

The country name, as in the original dataset, with minimal modification. Use extended_country_name instead if you want a consistent name.

anckar_ccode

The codebook says this is the COW code; it looks like a modified version of the Polity code (with the USSR = 364, for example). Use cown or GWn instead.

abbreviation

World Bank country abbreviation.

year

The calendar year.

democracy

This should be identical to version 2 of Boix, Miller and Rosato's Dichotomous democracy measure (democracy_omitteddata; see bmr). Supplemented for all countries for the period 2011-2016, as well as for Liechtenstein 1866-1990, Monaco 1862-2016, and San Marino 1800-1992, by Anckar and Fredriksson. 1 = democracy, 0 = non-democracy.

monarchy

Distinguishes between monarchies and republics in democratic countries. It is NA for non-democracies. Vales: 0 Republic, 1 Monarchy, 7 Hybrid.

regimebroadcat

Regime type. Classification with broad categories: Parliamentarism, Semi-presidentialism, Presidentialism, Semi-monarchy, Party-based rule, Personalist rule, Military rule, Absolute monarchy, Oligarchy

regimenarrowcat

Regime type. Classification with narrow categories: Parliamentarism, Semi-presidentialism, Presidentialism, Semi-monarchy, Single-party rule, Multi-party authoritarian rule, Personalist rule, Military rule, Absolute monarchy, Monarchic oligarchy, Other oligarchy.

popelection

Indicates if the head of state is popularly elected or not in democratic republics. Values 0 Head of state not popularly elected; 1 Head of state popularly elected. In a few cases this variable is not NA for non-democracies; it is unclear why.

Standard descriptive variables (generated by this package)

extended_country_name

The name of the country in the Gleditsch-Ward system of states, or the official name of the entity (for non-sovereign entities and states not in the Gleditsch and Ward system of states) or else a common name for disputed cases that do not have an official name (e.g., Western Sahara, Hyderabad). The Gleditsch and Ward scheme sometimes indicates the common name of the country and (in parentheses) the name of an earlier incarnation of the state: thus, they have Germany (Prussia), Russia (Soviet Union), Madagascar (Malagasy), etc. For details, see Gleditsch, Kristian S. & Michael D. Ward. 1999. "Interstate System Membership: A Revised List of the Independent States since 1816." International Interactions 25: 393-413. The list can be found at http://privatewww.essex.ac.uk/~ksg/statelist.html.

GWn

Gleditsch and Ward's numeric country code, from the Gleditsch and Ward list of independent states.

cown

The Correlates of War numeric country code, 2016 version. This differs from Gleditsch and Ward's numeric country code in a few cases. See http://www.correlatesofwar.org/data-sets/state-system-membership for the full list.

in_GW_system

Whether the state is "in system" (that is, is independent and sovereign), according to Gleditsch and Ward, for this particular date. Matches at the end of the year; so, for example South Vietnam 1975 is FALSE because, according to Gleditsch and Ward, the country ended on April 1975 (being absorbed by North Vietnam). It is also TRUE for dates beyond 2012 for countries that did not end by then, depsite the fact that the Gleditsch and Ward list has not been updated since.

See also