Boix-Miller-Rosato dichotomous coding of democracy, 1800-2020, version 4.0

A dataset containing the Boix-Miller-Rosato (BMR) dichotomous coding of democracy, first used in the paper C. Boix, M. Miller, and S. Rosato. "A Complete Data Set of Political Regimes, 1800–2007". In: Comparative Political Studies 46.12 (2012), pp. 1523-1554. DOI: 10.1177/0010414012463905. The complete dataset can be downloaded at https://sites.google.com/site/mkmtwo/data. The complete codebook is available there as well.

Usage

bmr

Format

An object of class tbl_df (inherits from tbl, data.frame) with 19775 rows and 16 columns.

Source

C. Boix, M. Miller, and S. Rosato. "A Complete Data Set of Political Regimes, 1800–2007". In: Comparative Political Studies 46.12 (2012), pp. 1523-1554. DOI: 10.1177/0010414012463905. Available at https://dataverse.harvard.edu/dataset.xhtml?persistentId=doi:10.7910/DVN/FJLMKT (versions 1-3) and https://sites.google.com/site/mkmtwo/data (version 4.0).

Details

bmr_country

The country name in BMR.

bmr_ccode

Numeric polity code code in BMR.

abbreviation

World Bank 3-letter code.

abbreviation_undp

UNDP 3-letter code.

year

Year

democracy

Dichotomous democracy measure. 1 = Democracy.

sovereign

Dichotomous indicator of sovereignty/independence (if 0, democracy is NA).

democracy_trans

-1 if democratic breakdown, 0 if no change, 1 if democratic transition.

democracy_breakdowns

Previous number of democratic breakdowns.

democracy_duration

Consecutive years of current regime type.

democracy_omitteddata

This is the same measure as democracy, except it records an NA for countries occupied during an international war (e.g., the Netherlands 1940-44) or experiencing state collapse during a civil war (e.g., Lebanon 1976-89). The democracy variable instead fills in these years as continuations of the same regime type.

democracy_femalesuffrage

This adjusts democracy by also requiring that at least half of adult women have the right to vote. 30 countries change values.

Standard descriptive variables (generated by this package)

extended_country_name

The name of the country in the Gleditsch-Ward system of states, or the official name of the entity (for non-sovereign entities and states not in the Gleditsch and Ward system of states) or else a common name for disputed cases that do not have an official name (e.g., Western Sahara, Hyderabad). The Gleditsch and Ward scheme sometimes indicates the common name of the country and (in parentheses) the name of an earlier incarnation of the state: thus, they have Germany (Prussia), Russia (Soviet Union), Madagascar (Malagasy), etc. For details, see Gleditsch, Kristian S. & Michael D. Ward. 1999. "Interstate System Membership: A Revised List of the Independent States since 1816." International Interactions 25: 393-413. The list can be found at http://privatewww.essex.ac.uk/~ksg/statelist.html.

GWn

Gleditsch and Ward's numeric country code, from the Gleditsch and Ward list of independent states.

cown

The Correlates of War numeric country code, 2016 version. This differs from Gleditsch and Ward's numeric country code in a few cases. See http://www.correlatesofwar.org/data-sets/state-system-membership for the full list.

in_GW_system

Whether the state is "in system" (that is, is independent and sovereign), according to Gleditsch and Ward, for this particular date. Matches at the end of the year; so, for example South Vietnam 1975 is FALSE because, according to Gleditsch and Ward, the country ended on April 1975 (being absorbed by North Vietnam). It is also TRUE for dates beyond 2012 for countries that did not end by then, depsite the fact that the Gleditsch and Ward list has not been updated since.

See also