Create a panel of countries belonging in the state system.

Create a panel of countries belonging in the state system.

create_panel(system = c("cow", "GW", "polity"), max_year = NULL)

Arguments

system

The state system to target. The possibilities are "cow" (Correlates of War, the default), "GW" (Gleditsch and Ward, from Gleditsch, Kristian S. & Michael D. Ward. 1999. "Interstate System Membership: A Revised List of the Independent States since 1816." International Interactions 25: 393-413, available at http://privatewww.essex.ac.uk/~ksg/statelist.html), or "polity" (for the Polity IV system of states).

max_year

The maximum year in the panel. If NULL, uses the current year.

Value

A tibble with the country code in the relevant system of states, the official country name and years when the country belongs to the system, and the relevant start and end dates for entry and exit into the system.

Examples

create_panel("cow") # The correlates of War system is the default
#> # A tibble: 17,036 x 5
#>     cown cow_country_name         cow_startdate cow_enddate  year
#>    <int> <chr>                    <date>        <date>      <dbl>
#>  1     2 United States of America 1816-01-01    NA           1816
#>  2     2 United States of America 1816-01-01    NA           1817
#>  3     2 United States of America 1816-01-01    NA           1818
#>  4     2 United States of America 1816-01-01    NA           1819
#>  5     2 United States of America 1816-01-01    NA           1820
#>  6     2 United States of America 1816-01-01    NA           1821
#>  7     2 United States of America 1816-01-01    NA           1822
#>  8     2 United States of America 1816-01-01    NA           1823
#>  9     2 United States of America 1816-01-01    NA           1824
#> 10     2 United States of America 1816-01-01    NA           1825
#> # ... with 17,026 more rows
create_panel("GW") # Use this for the Gleditsch and Ward system
#> # A tibble: 19,938 x 5
#>      GWn GW_country_name          GW_startdate GW_enddate  year
#>    <dbl> <chr>                    <date>       <date>     <dbl>
#>  1     2 United States of America 1816-01-01   NA          1816
#>  2     2 United States of America 1816-01-01   NA          1817
#>  3     2 United States of America 1816-01-01   NA          1818
#>  4     2 United States of America 1816-01-01   NA          1819
#>  5     2 United States of America 1816-01-01   NA          1820
#>  6     2 United States of America 1816-01-01   NA          1821
#>  7     2 United States of America 1816-01-01   NA          1822
#>  8     2 United States of America 1816-01-01   NA          1823
#>  9     2 United States of America 1816-01-01   NA          1824
#> 10     2 United States of America 1816-01-01   NA          1825
#> # ... with 19,928 more rows
create_panel("polity", max_year = 1997) # You can specify a maximum year
#> # A tibble: 13,964 x 5
#>    polity_ccode polity_country_name polity_startdate polity_enddate  year
#>           <dbl> <chr>               <date>           <date>         <dbl>
#>  1            2 United States       1800-01-01       NA              1800
#>  2            2 United States       1800-01-01       NA              1801
#>  3            2 United States       1800-01-01       NA              1802
#>  4            2 United States       1800-01-01       NA              1803
#>  5            2 United States       1800-01-01       NA              1804
#>  6            2 United States       1800-01-01       NA              1805
#>  7            2 United States       1800-01-01       NA              1806
#>  8            2 United States       1800-01-01       NA              1807
#>  9            2 United States       1800-01-01       NA              1808
#> 10            2 United States       1800-01-01       NA              1809
#> # ... with 13,954 more rows