Extract cutpoints from a UD model in a tidy format.

This function takes a model of the democracy scores and extracts the discrimination parameters, score cutpoints, and standard errors for all the variables involved, putting these into a tidy data frame.

cutpoints(model, type = "score")

Arguments

model
A mirt `SingleGroupClass-class` model of the democracy scores.

type
A string specifying the cutpoint type. Can be (an abbreviation of) "score" (for score cutpoints) or "discrimination" (for discrimination parameters). Default is "score."

Value

A tibble with either score cutpoints for each variable used to construct the latent scores in terms of the latent variable (the default), or discrimination parameters for each variable used to construct the index. For the score cutpoints (type = 'score'), the columns estimate, pct975, and pct025 report the IRT parametrization of the model estimates, a normalized measure in the same scale as the latent variable.

Examples

# \donttest{
# Replicate the official UDS 2011 release and calculate its cutpoints
library(dplyr)
library(mirt)

#> Loading required package: stats4

#> Loading required package: lattice

democracy_data <- generate_democracy_scores_dataset(selection = "_pmm", output_format = "wide")

#> Adding fh_pmm data

#> Adding polity_pmm data

#> Adding Arat data

#> Adding blm_pmm data

#> Adding Bollen data

#> Adding Hadenius data

#> Adding mainwaring_pmm data

#> Adding Munck data

#> Adding pacl_pmm data

#> Adding polyarchy_pmm data

#> Adding prc_pmm data

#> Adding Vanhanen_pmm data

#> Finalizing

democracy_data <- prepare_democracy_data(democracy_data)

replication_2011_model <- mirt(democracy_data %>% select(matches("pmm")), model = 1, itemtype = "graded", SE = TRUE, verbose = FALSE)

cutpoints(replication_2011_model)# }

#> # A tibble: 85 x 6
#>    variable     estimate  pct025  pct975     se num_obs
#>    <chr>           <dbl>   <dbl>   <dbl>  <dbl>   <int>
#>  1 pmm_vanhanen  -0.163  -0.186  -0.137  0.0129    8965
#>  2 pmm_vanhanen   0.138   0.104   0.176  0.0191    8965
#>  3 pmm_vanhanen   0.424   0.377   0.475  0.0259    8965
#>  4 pmm_vanhanen   0.750   0.687   0.818  0.0347    8965
#>  5 pmm_vanhanen   1.01    0.931   1.09   0.0423    8965
#>  6 pmm_vanhanen   1.31    1.22    1.41   0.0517    8965
#>  7 pmm_vanhanen   1.74    1.62    1.87   0.0666    8965
#>  8 pmm_prc       -0.133  -0.156  -0.107  0.0134    6002
#>  9 pmm_prc       -0.0609 -0.0877 -0.0305 0.0155    6002
#> 10 pmm_prc        0.312   0.261   0.371  0.0301    6002
#> # ... with 75 more rows