Basic information about existing democracy measures

This data frame contains some basic info about all the democracy measures I know about.

democracy_info

Format

An object of class spec_tbl_df (inherits from tbl_df, tbl, data.frame) with 40 rows and 12 columns.

Variables

dataset

The basic name of the dataset. This usually matches the name of the dataset included in this package, except for datasets with more than one version (e.g., gwf matches gwf_all, gwf_all_extended, gwf_autocratic, and gwf_autocratic_extended). It also includes datasets not included in this package (e.g., vdem, accessible via the vdem package on Github (http://github.com/xmarquez/vdem), and polity IV, accesible via the download_polity_annual function in this package).

long_name

Longer name of the dataset, or official name, if any.

main_democracy_measure_col

The column or columns of the dataset containing a measure of democracy. There may be more than one; refer to the individual dataset documentation for details. If more than one measure, these are separated by commas.

measure_type

The type of measure. It can be dichotomous (disintguishing only between democracy and non-democracy, with perhaps special codes for state interruptions or no-authority conditions), trichotomous (including a semidemocracy or hybrid category, but no other distinctions), ordinal (including several levels of democracy), or continuous.

based_on

Some measures of democracy modify or build on existing ones. Many are based on Polity data, for example. This column lists the basis for the measures of democracy in each datasets, as far as I know base don my reading of the sources. In some cases no relationship to other measures could be identified; this does not mean that no relationship exists. In other cases a measure may draw on more than one existing dataset. (For example, Wahman-Teorell-Hadenius use both Freedom House and Polity to help them identify democracies).

in_pmm_replication

Logical. Lists whether the measure is found in Pemstein, Meserve, and Melton's (2013) replication data for their 2010 article introducing the Unified Democracy Scores (Pemstein, Meserve, and Melton 2010). In some cases this is the only electronic source I have been able to identify for the included measure (e.g., arat_pmm). It's also worth noting that in a few cases (e.g., prc_gasiorowski, prc_pmm) there are some identifiable differences between the original data and the data in the replication dataset. See the vignette in this package for further discussion.

categorical_regime_types

Whether the dataset contains information about categorical regime types - primarily forms of authoritarianism (e.g., monarchies and military regimes).

user_extendable

Whether the dataset is extendable in time by the user. Some datasets (e.g., gwf, ulfelder, bnr, magaloni) contain information about regime duration that can be used to extend their range back in time; others are latent variable indexes that can be extended using specific packages (e.g., uds, which can be extended using the QuickUDS package on Github (http://github.com/xmarquez/QuickUDS). The "extended" versions of some datasets are included in this package.

donwloadable

Whether this package includes a function to download the dataset directly from the original source.

included_in_package

Whether the dataset is included in this package. Some are not, either due to use restrictions, or because they are updated with some frequency (e.g., fh), or because they are too big (e.g., vdem, which is accessible via the vdem package on Github (http://github.com/xmarquez/vdem).

first_published_use

Year of first published use (as far as I know). In some cases the data used in the first published use is very different from the current data; for example, as far as I know the Polity I data was first used in a 1975 study (though its compilation started in the late 1960s), but the current Polity IV data is very different.

notes

Other notes. Evolving.

References

Pemstein, Daniel, Stephen Meserve, and James Melton. 2010. Democratic Compromise: A Latent Variable Analysis of Ten Measures of Regime Type. Political Analysis 18 (4): 426-449.

Pemstein, Daniel, Stephen A. Meserve, and James Melton. 2013. "Replication data for: Democratic Compromise: A Latent Variable Analysis of Ten Measures of Regime Type." In: Harvard Dataverse. http://hdl.handle.net/1902.1/PMM