Renske Doorenspleet's Democracy Dataset

A dataset containing the country-year version of the democracy classification described in Doorenspleet, Renske. 2000. Reassessing the Three Waves of Democratization. World Politics 52 (03): 384-406. DOI: 10.1017/S0043887100016580. http://dx.doi.org/10.1017/S0043887100016580.

doorenspleet

Format

An object of class tbl_df (inherits from tbl, data.frame) with 13877 rows and 10 columns.

Source

Doorenspleet, Renske. 2000. Reassessing the Three Waves of Democratization. World Politics 52 (03): 384-406. DOI: 10.1017/S0043887100016580. http://dx.doi.org/10.1017/S0043887100016580.

Details

doorenspleet_country

Country name as in Doorenspleet.

doorenspleet_regime

Regime type: A = Authoritarian, D = Democracy, I = interruption.

doorenspleet

Regime type: 1 = Authoritarian, 2 = Democracy, interruption codes as NA.

year

Year. Regimes are coded as of the 31 Dec of the year.

start_year

First year of the regime.

end_year

Last year of the regime.

Standard descriptive variables (generated by this package)

extended_country_name

The name of the country in the Gleditsch-Ward system of states, or the official name of the entity (for non-sovereign entities and states not in the Gleditsch and Ward system of states) or else a common name for disputed cases that do not have an official name (e.g., Western Sahara, Hyderabad). The Gleditsch and Ward scheme sometimes indicates the common name of the country and (in parentheses) the name of an earlier incarnation of the state: thus, they have Germany (Prussia), Russia (Soviet Union), Madagascar (Malagasy), etc. For details, see Gleditsch, Kristian S. & Michael D. Ward. 1999. "Interstate System Membership: A Revised List of the Independent States since 1816." International Interactions 25: 393-413. The list can be found at http://privatewww.essex.ac.uk/~ksg/statelist.html.

GWn

Gleditsch and Ward's numeric country code, from the Gleditsch and Ward list of independent states.

cown

The Correlates of War numeric country code, 2016 version. This differs from Gleditsch and Ward's numeric country code in a few cases. See http://www.correlatesofwar.org/data-sets/state-system-membership for the full list.

in_GW_system

Whether the state is "in system" (that is, is independent and sovereign), according to Gleditsch and Ward, for this particular date. Matches at the end of the year; so, for example South Vietnam 1975 is FALSE because, according to Gleditsch and Ward, the country ended on April 1975 (being absorbed by North Vietnam). It is also TRUE for dates beyond 2012 for countries that did not end by then, depsite the fact that the Gleditsch and Ward list has not been updated since.

See also