Freedom House "Freedom in the World" data

Downloads the 2023 update of the Freedom House "Freedom in the World" data and processes it using country_year_coder. The original data is available at https://freedomhouse.org/report-types/freedom-world

download_fh(
  url,
  include_territories = FALSE,
  verbose = TRUE,
  return_raw = FALSE,
  ...
)

Source

The "Freedom in the World" dataset from Freedom House, updated to 2022 (Freedom in the World 2023 Report). Original data and methodology is available at https://freedomhouse.org/report/freedom-world

Arguments

url

The URL of the dataset. Defaults to https://freedomhouse.org/sites/default/files/2023-02/Country_and_Territory_Ratings_and_Statuses_FIW_1973-2023%20.xlsx

include_territories

Whether to include scores from non-independent territories (e.g., Indian Kashmir, Northern Ireland) compiled by FH. Default is FALSE.

verbose

Whether to print a running commentary of what the function is doing while processing the data.

return_raw

Whether to return the raw data, without any processing. Default is FALSE.

...

Other parameters passed to country_year_coder.

Value

A time-series tidy version of the FH dataset, with the following variables:

country: The original country name.

year: The calendar year. This is approximate. The surveys cover specific periods in the original data that do not always overlap with a single year. In particular, the year 1981 is "skipped" -- a single survey covers Jan. 1981 - Aug. 1982 and its value is assigned to 1982 here. Surveys from 1981 to 1988 covered the periods Jan. 1981-Aug. 1982, Aug. 1982-Nov. 1983, Nov. 1983-Nov. 1984, Nov. 1984-Nov. 1985, Nov. 1985-Nov. 1986, Nov. 1986-Nov. 1987, Nov. 1987-Nov. 1988, and Nov.1988-Dec.1989, and results are assigned to years 1982, 1983, 1984, 1985, 1986, 1987, 1988, and 1989.

pr: Political rights, from 1 to 7, where 7 means more violations of political rights. Note that South Africa in 1973 was assigned two values (one for white and one for black South Africa). Here it is given the "not free" values. See notes for details.

cl: Civil liberties, from 1 to 7, where 7 means more violations of civil rights. Note that South Africa in 1973 was assigned two values (one for white and one for black South Africa). Here it is given the "not free" values. See notes for details

status: Political status. Free, not free, or partly free. See notes for details.

fh_total: Sum of pr and cl. Higher values are "less free."

fh_total_reversed: Reversed fh_total, so that higher values are "more free."

Standard descriptive variables (generated by this package)

extended_country_name

The name of the country in the Gleditsch-Ward system of states, or the official name of the entity (for non-sovereign entities and states not in the Gleditsch and Ward system of states) or else a common name for disputed cases that do not have an official name (e.g., Western Sahara, Hyderabad). The Gleditsch and Ward scheme sometimes indicates the common name of the country and (in parentheses) the name of an earlier incarnation of the state: thus, they have Germany (Prussia), Russia (Soviet Union), Madagascar (Malagasy), etc. For details, see Gleditsch, Kristian S. & Michael D. Ward. 1999. "Interstate System Membership: A Revised List of the Independent States since 1816." International Interactions 25: 393-413. The list can be found at http://privatewww.essex.ac.uk/~ksg/statelist.html.

GWn

Gleditsch and Ward's numeric country code, from the Gleditsch and Ward list of independent states.

cown

The Correlates of War numeric country code, 2016 version. This differs from Gleditsch and Ward's numeric country code in a few cases. See http://www.correlatesofwar.org/data-sets/state-system-membership for the full list.

in_GW_system

Whether the state is "in system" (that is, is independent and sovereign), according to Gleditsch and Ward, for this particular date. Matches at the end of the year; so, for example South Vietnam 1975 is FALSE because, according to Gleditsch and Ward, the country ended on April 1975 (being absorbed by North Vietnam). It is also TRUE for dates beyond 2012 for countries that did not end by then, depsite the fact that the Gleditsch and Ward list has not been updated since.

Methodological notes

From FH's Freedom in the World methodology discussion:

Methodological changes have been effected periodically. For discussions of these changes, please consult the methodology essays for various editions of the survey.

For a full explanation of the current methodology, please consult the most recent edition of the survey.

"PR" stands for "Political Rights," "CL" stands for "Civil Liberties," and "Status" is the Freedom Status. Political Rights and Civil Liberties are measured on a one-to-seven scale, with one representing the highest degree of Freedom and seven the lowest.

Until 2003, countries whose combined average ratings for Political Rights and for Civil Liberties fell between 1.0 and 2.5 were designated "Free"; between 3.0 and 5.5 "Partly Free," and between 5.5 and 7.0 "Not Free." Beginning with the ratings for 2003, countries whose combined average ratings fall between 3.0 and 5.0 are "Partly Free," and those between 5.5 and 7.0 are "Not Free." Ratings for territories are not included in this table.

Several countries became independent, split into two or more countries, or merged with a neighboring state. Scores for these countries are given only for the period of their existence as independent states.

The survey years refer to the various editions of Freedom in the World, except for Jan.-Feb. 1973 through Jan.-Feb. 1977, which are from the bimonthly journal Freedom at Issue. The first Freedom in the World book-length survey was the 1978 edition.

For 1972, South Africa was rated as "White" (2,3 Free) and "Black" (5,6 Not Free). For Yugoslavia, ratings from 1999 to 2002 were for the country that remained following the departures between 1991 and 1992 of Slovenia, Croatia, Macedonia and Bosnia-Herzegovina. In February 2003, the Yugoslav parliament adopted a constitutional charter establishing the state of Serbia and Montenegro. Thus, beginning in 2003, Yugoslavia is listed as "Serbia and Montenegro." The State Union of Serbia and Montenegro dissolved when Montenegro withdrew in June 2006, making Serbia an independent state. Thus, the ratings for Serbia and Montenegro are listed separately beginning in 2006. Kosovo was first listed as a territory beginning in 1992. Since 2009, it is listed as an independent country. South Sudan was first listed as an independent country in 2011 after officially separating from Sudan in July 2011. The former Zaire is listed under Congo (Kinshasa), and the former Western Samoa is listed under Samoa.

Note

The datasets downloaded by the download_* family of functions are not directly available in this package. You will always need to directly download them in order to use them.

See also

Examples

if (FALSE) {
fh <- download_fh()
fh
fh_with_territories <- download_fh(include_territories = TRUE)
fh_with_territories
}