Downloads the 2017 update of the Freedom House Electoral Democracies list and processes it using country_year_coder.

The original data is available at https://freedomhouse.org/sites/default/files/FIW2017_Data.zip

download_fh_electoral(url, verbose = TRUE, return_raw = FALSE, ...)

Arguments

url
The URL of the dataset. Defaults to https://freedomhouse.org/sites/default/files/FIW2017_Data.zip

verbose
Whether to print a running commentary of what the function is doing while processing the data.

return_raw
Whether to return the raw data, without any processing. Default is FALSE.

...
Other parameters passed to country_year_coder.

Source

The "Freedom in the World" dataset from Freedom House, updated to 2016. Original data and methodology is available at https://freedomhouse.org/report/freedom-world/freedom-world-2017

Value

A time-series tidy version of the FH electoral democracies dataset, with the following variables:

country: The original country name.

year: The calendar year. This is approximate. The surveys cover specific periods in the original data that do not always overlap with a single year. In particular, the year 1981 is "skipped" – a single survey covers Jan. 1981 - Aug. 1982 and its value is assigned to 1982 here.

electoral: Electoral democracy indicator. TRUE = yes, FALSE = no.

Standard descriptive variables (generated by this package)

extended_country_name

The name of the country in the Gleditsch-Ward system of states, or the official name of the entity (for non-sovereign entities and states not in the Gleditsch and Ward system of states) or else a common name for disputed cases that do not have an official name (e.g., Western Sahara, Hyderabad). The Gleditsch and Ward scheme sometimes indicates the common name of the country and (in parentheses) the name of an earlier incarnation of the state: thus, they have Germany (Prussia), Russia (Soviet Union), Madagascar (Malagasy), etc. For details, see Gleditsch, Kristian S. & Michael D. Ward. 1999. "Interstate System Membership: A Revised List of the Independent States since 1816." International Interactions 25: 393-413. The list can be found at http://privatewww.essex.ac.uk/~ksg/statelist.html.

GWn

Gleditsch and Ward's numeric country code, from the Gleditsch and Ward list of independent states.

cown

The Correlates of War numeric country code, 2016 version. This differs from Gleditsch and Ward's numeric country code in a few cases. See http://www.correlatesofwar.org/data-sets/state-system-membership for the full list.

in_GW_system

Whether the state is "in system" (that is, is independent and sovereign), according to Gleditsch and Ward, for this particular date. Matches at the end of the year; so, for example South Vietnam 1975 is FALSE because, according to Gleditsch and Ward, the country ended on April 1975 (being absorbed by North Vietnam). It is also TRUE for dates beyond 2012 for countries that did not end by then, depsite the fact that the Gleditsch and Ward list has not been updated since.

Note

The datasets downloaded by the download_* family of functions are not directly available in this package. You will always need to directly download them in order to use them.

See also

Other democracy: LIED, PIPE, arat_pmm, blm, bmr, bnr, bollen_pmm, doorenspleet, download_fh, download_reign, download_wgi_voice_and_accountability, fh_pmm, gwf_all, hadenius_pmm, kailitz, magaloni, mainwaring, munck_pmm, pacl, peps, pitf, polity_pmm, polyarchy_dimensions, polyarchy, prc_gasiorowski, svolik_regime, uds_2014, ulfelder, utip, vanhanen, wahman_teorell_hadenius

Other Freedom House: download_fh, fh_pmm

Other ordinal democracy indexes: download_fh, fh_pmm, polity_pmm

Examples

# NOT RUN {
fh <- download_fh_electoral()
fh
# }