Downloads the 2020 update of the Freedom House "Freedom in the World" All Data 2013-2020 file and processes it using country_year_coder.

The original data is available at https://freedomhouse.org/report-types/freedom-world

download_fh_full(url, verbose = TRUE, return_raw = FALSE, ...)

Arguments

url
The URL of the dataset. Defaults to https://freedomhouse.org/sites/default/files/2020-02/2020_All_Data_FIW_2013-2020.xlsx

verbose
Whether to print a running commentary of what the function is doing while processing the data.

return_raw
Whether to return the raw data, without any processing. Default is FALSE.

...
Other parameters passed to country_year_coder.

Source

The "Freedom in the World" dataset from Freedom House, updated to 2019 (Freedom in the World 2020 Report). Original data and methodology is available at https://freedomhouse.org/report/freedom-world

Value

A time-series tidy version of the FH "all data 2013-2020" dataset, with the following variables:

country: The original country name.

region: Freedom House's regional classification.

country_or_territory: Whether Freedom House classifies the unit as a country or a territory.

edition: the edition of FIW in which the data appeared.

status: Freedom status: Free, Partly Free, Not Free.

year: The calendar year of the survey. (Added).

pr: Political rights, from 1 to 7, where 7 means more violations of political rights.

cl: Civil liberties, from 1 to 7, where 7 means more violations of civil rights.

A1: score for electoral process subcategory question 1.

A2: score for electoral process subcategory question 2.

A3: score for electoral process subcategory question 3.

A: aggregate score for electoral process subcategory.

B1: score for political pluralism and participation subcategory question

B2: score for political pluralism and participation subcategory question

B3: score for political pluralism and participation subcategory question

B4: score for political pluralism and participation subcategory question

B: aggregate score for political pluralism and participation subcategory.

C1: score for functioning of government subcategory question 1.

C2: score for functioning of government subcategory question 2.

C3: score for functioning of government subcategory question 3.

C: aggregate score for functioning of government subcategory.

AddQ: score for Additional Discretionary Question (B). Note in FIW2018, AddA (Additional Discretionary Question A) was eliminated, and AddB (Additional Discretionary Question B) was renamed AddQ.

AddA: score for Additional Discretionary Question (B). Note in FIW2018, AddA (Additional Discretionary Question A) was eliminated, and AddB (Additional Discretionary Question B) was renamed AddQ.

pr_full: aggregate score for subsections A-C on political rights.

D1: score for freedom of expression subcategory question 1.

D2: score for freedom of expression subcategory question 2.

D3: score for freedom of expression subcategory question 3.

D4: score for freedom of expression subcategory question 4.

D: aggregate score for freedom of expression subcategory.

E1: score for associational and organizational rights subcategory question

E2: score for associational and organizational rights subcategory question

E3: score for associational and organizational rights subcategory question

E: aggregate score for associational and organizational rights subcategory.

F1: score for rule of law subcategory question 1.

F2: score for rule of law subcategory question 2.

F3: score for rule of law subcategory question 3.

F4: score for rule of law subcategory question 4.

F: aggregate score for rule of law subcategory.

G1: score for personal autonomy and individual rights subcategory question 1.

G2: score for personal autonomy and individual rights question 2.

G3: score for personal autonomy and individual rights subcategory question 3.

G4: score for personal autonomy and individual rights subcategory question 4.

G: aggregate score for personal autonomy and individual rights subcategory.

cl_full: aggregate score for subsections D-G on civil liberties.

total: full score (pr_full + cl_full), out of 100.

Standard descriptive variables (generated by this package)

extended_country_name

The name of the country in the Gleditsch-Ward system of states, or the official name of the entity (for non-sovereign entities and states not in the Gleditsch and Ward system of states) or else a common name for disputed cases that do not have an official name (e.g., Western Sahara, Hyderabad). The Gleditsch and Ward scheme sometimes indicates the common name of the country and (in parentheses) the name of an earlier incarnation of the state: thus, they have Germany (Prussia), Russia (Soviet Union), Madagascar (Malagasy), etc. For details, see Gleditsch, Kristian S. & Michael D. Ward. 1999. "Interstate System Membership: A Revised List of the Independent States since 1816." International Interactions 25: 393-413. The list can be found at http://privatewww.essex.ac.uk/~ksg/statelist.html.

GWn

Gleditsch and Ward's numeric country code, from the Gleditsch and Ward list of independent states.

cown

The Correlates of War numeric country code, 2016 version. This differs from Gleditsch and Ward's numeric country code in a few cases. See http://www.correlatesofwar.org/data-sets/state-system-membership for the full list.

in_GW_system

Whether the state is "in system" (that is, is independent and sovereign), according to Gleditsch and Ward, for this particular date. Matches at the end of the year; so, for example South Vietnam 1975 is FALSE because, according to Gleditsch and Ward, the country ended on April 1975 (being absorbed by North Vietnam). It is also TRUE for dates beyond 2012 for countries that did not end by then, depsite the fact that the Gleditsch and Ward list has not been updated since.

Methodological notes

From FH's Freedom in the World methodology discussion:

Methodological changes have been effected periodically. For discussions of these changes, please consult the methodology essays for various editions of the survey.

For a full explanation of the current methodology, please consult the most recent edition of the survey.

"PR" stands for "Political Rights," "CL" stands for "Civil Liberties," and "Status" is the Freedom Status. Political Rights and Civil Liberties are measured on a one-to-seven scale, with one representing the highest degree of Freedom and seven the lowest.

Until 2003, countries whose combined average ratings for Political Rights and for Civil Liberties fell between 1.0 and 2.5 were designated "Free"; between 3.0 and 5.5 "Partly Free," and between 5.5 and 7.0 "Not Free." Beginning with the ratings for 2003, countries whose combined average ratings fall between 3.0 and 5.0 are "Partly Free," and those between 5.5 and 7.0 are "Not Free."

Several countries became independent, split into two or more countries, or merged with a neighboring state. Scores for these countries are given only for the period of their existence as independent states.

Note

The datasets downloaded by the download_* family of functions are not directly available in this package. You will always need to directly download them in order to use them.

See also

Examples

if (FALSE) {
fh_full <- download_fh_full()
fh_full
}