Downloads the 2017 update of the Polity IV dataset (annual time series) and processes it using country_year_coder.

The original data is available at http://www.systemicpeace.org/inscrdata.html

download_polity_annual(url, verbose = TRUE, return_raw = FALSE, ...)

Arguments

url
The URL of the dataset. Defaults to http://www.systemicpeace.org/inscr/p4v2016.sav

verbose
Whether to print a running commentary of what the function is doing while processing the data.

return_raw
Whether to return the raw data, without any processing. Default is FALSE.

...
Other parameters passed to country_year_coder.

Value

The annual time series version of the polity dataset, as a tibble, with the additional columns produced by country_year_coder. Consult the Polity IV codebook for further description.

Standard descriptive variables (generated by this package)

extended_country_name

The name of the country in the Gleditsch-Ward system of states, or the official name of the entity (for non-sovereign entities and states not in the Gleditsch and Ward system of states) or else a common name for disputed cases that do not have an official name (e.g., Western Sahara, Hyderabad). The Gleditsch and Ward scheme sometimes indicates the common name of the country and (in parentheses) the name of an earlier incarnation of the state: thus, they have Germany (Prussia), Russia (Soviet Union), Madagascar (Malagasy), etc. For details, see Gleditsch, Kristian S. & Michael D. Ward. 1999. "Interstate System Membership: A Revised List of the Independent States since 1816." International Interactions 25: 393-413. The list can be found at http://privatewww.essex.ac.uk/~ksg/statelist.html.

GWn

Gleditsch and Ward's numeric country code, from the Gleditsch and Ward list of independent states.

cown

The Correlates of War numeric country code, 2016 version. This differs from Gleditsch and Ward's numeric country code in a few cases. See http://www.correlatesofwar.org/data-sets/state-system-membership for the full list.

in_GW_system

Whether the state is "in system" (that is, is independent and sovereign), according to Gleditsch and Ward, for this particular date. Matches at the end of the year; so, for example South Vietnam 1975 is FALSE because, according to Gleditsch and Ward, the country ended on April 1975 (being absorbed by North Vietnam). It is also TRUE for dates beyond 2012 for countries that did not end by then, depsite the fact that the Gleditsch and Ward list has not been updated since.

Note

The datasets downloaded by the download_* family of functions are not directly available in this package. You will always need to directly download them in order to use them.

See also

polity_pmm

Examples

# NOT RUN {
polity_annual <- download_polity_annual()
polity_annual
# }