Democracy, Voice, and Accountability Index from the World Governance Indicators

Download the latest version (data up to 2021) of the Democracy, Voice, and Accountability index from the World Governance Indicators and process it via country_year_coder

download_wgi_voice_and_accountability(
  url,
  verbose = TRUE,
  return_raw = FALSE,
  ...
)

Source

Worldwide Governance Indicators (http://www.govindicators.org).

Arguments

url

The URL of the dataset. Defaults to http://info.worldbank.org/governance/wgi/Home/downLoadFile?fileName=wgidataset.xlsx

verbose

Whether to print a running commentary of what the function is doing while processing the data.

return_raw

Whether to return the raw data, without any processing. Default is FALSE.

...

Other parameters passed to country_year_coder.

Value

A dataset containing the country-year version of the latest update (to 2021) of the World Governance Indicators "Democracy, Voice and Accountability" Index, compiled by Daniel Kaufmann and Aart Kraay. The original data are available at http://info.worldbank.org/governance/wgi/index.aspx#home, along with a full explanation of the methodology used to compile the index.

According to the compilers, "Voice and accountability captures perceptions of the extent to which a country's citizens are able to participate in selecting their government, as well as freedom of expression, freedom of association, and a free media."

wb_country

Country name (World Bank standard).

wb_code

World bank 3 character country code.

year

Year.

Estimate

Estimate of governance (ranges from approximately -2.5 (weak) to 2.5 (strong) governance performance), with a mean of 0.

StdErr

Standard error reflects variability around the point estimate of governance.

NumSrc

Number of data sources on which estimate is based. The methodology document lists, among these, the EIU Democracy Index, Freedom House, the World Economic Forum Global Competitiveness Report, the Gallup World Poll, the Institutional Profiles Database, the Political Risk Services International Country Risk Guide, and Reporters Without Borders Press Freedom Index. See http://info.worldbank.org/governance/wgi/va.pdf for more details.

Rank

Percentile rank among all countries (ranges from 0 (lowest) to 100 (highest) rank).

Lower

Lower bound of 90% confidence interval for governance, in percentile rank terms.

Upper

Upper bound of 90% confidence interval for governance, in percentile rank terms

Standard descriptive variables (generated by this package)

extended_country_name

The name of the country in the Gleditsch-Ward system of states, or the official name of the entity (for non-sovereign entities and states not in the Gleditsch and Ward system of states) or else a common name for disputed cases that do not have an official name (e.g., Western Sahara, Hyderabad). The Gleditsch and Ward scheme sometimes indicates the common name of the country and (in parentheses) the name of an earlier incarnation of the state: thus, they have Germany (Prussia), Russia (Soviet Union), Madagascar (Malagasy), etc. For details, see Gleditsch, Kristian S. & Michael D. Ward. 1999. "Interstate System Membership: A Revised List of the Independent States since 1816." International Interactions 25: 393-413. The list can be found at http://privatewww.essex.ac.uk/~ksg/statelist.html.

GWn

Gleditsch and Ward's numeric country code, from the Gleditsch and Ward list of independent states.

cown

The Correlates of War numeric country code, 2016 version. This differs from Gleditsch and Ward's numeric country code in a few cases. See http://www.correlatesofwar.org/data-sets/state-system-membership for the full list.

in_GW_system

Whether the state is "in system" (that is, is independent and sovereign), according to Gleditsch and Ward, for this particular date. Matches at the end of the year; so, for example South Vietnam 1975 is FALSE because, according to Gleditsch and Ward, the country ended on April 1975 (being absorbed by North Vietnam). It is also TRUE for dates beyond 2012 for countries that did not end by then, depsite the fact that the Gleditsch and Ward list has not been updated since.

Note

The datasets downloaded by the download_* family of functions are not directly available in this package. You will always need to directly download them in order to use them.

See also

Examples

if (FALSE) {
download_wgi_voice_and_accountability()}