Extended UDS

A dataset extending the Unified Democracy Scores of Pemstein, Meserve, and Melton (2010) to the 19th century (and sometimes before), updating it with 2013-2022 data, and calculating scores for countries not in the official UD release. Cite both Pemstein, Meserve, and Melton 2010 as well as Marquez 2016.

extended_uds

Format

An object of class tbl_df (inherits from tbl, data.frame) with 36841 rows and 20 columns.

Overview

This dataset contains an extension of the Unified Democracy Scores of Pemstein, Meserve, and Melton (2010). PMM 2010 use a latent variable approach to combine diverse measurements of democracy for a broad panel of countries from 1946 to 2012; this dataset extends the scores by using more democracy measures with broader temporal and spatial coverage, generating democracy scores for 36841 country-years (274 distinct countries and non-sovereign territories). It is generated by the function generate_extended_uds.

The measurement of democracy is complicated and controversial. The extended UDS makes few judgments about what measures of democracy should be used to generate latent democracy scores, as long as they have been used in scholarly work. It uses dichotomous, trichotomous, ordinal, and continuous indices; indices that focus primarily on the "competition" and indices that focus on the "participation" dimension of democracy; "thick" indices that attempt to measure a wide variety of characteristics plausibly attributed to democracy; and "minimalist" indexes that restrict themselves to the bare minimum of competition. Nevertheless, most of these indexes agree that democracy has something to do with competition and participation, even if they weight these dimensions somewhat differently, and even if they include other things, such as civil rights.

The latent variable score is calculated using the democracy measures from the following datasets: anckar, anrr, arat_pmm, blm, bmr (democracy_femalesuffrage indicator only), bnr_extended, bti, bollen_pmm, svmdi (continuous indicator), doorenspleet, eiu, fh (fh_total_reversed indicator), fh_electoral, gwf_all_extended, hadenius_pmm, kailitz (trichotomous indicator of democracy, with electoral autocracy as the middle category), LIED, magaloni_extended, mainwaring, munck_pmm, pacl, pacl_update, peps (PEPS1v indicator), pitf (non-binary indicator), polity (polity 2 indicator), polyarchy (both contestation and participation dimensions), polyarchy_dimensions (both contestation and participation dimensions), prc, REIGN, svolik, ulfelder_extended, utip (trichotomous indicator, with one-party democracy as the middle category), the v2x_polyarchy indicator from vdem_simple, vanhanen (the vanhanen_democratization indicator only), wgi, and wahman_teorell_hadenius (the from the regimenyrobust variable). Please refer to the documentation of generate_democracy_scores_dataset for more details on these measures of democracy, and to Marquez 2016 for more details on the construction of the extended latent variable model used to generate the scores.

Variables

z1, z1_adj, z1_as_prob, z1_adj_as_prob

The latent democracy score. z1 is the "raw" (unadjusted) score, with mean approximately 0 and sd 1; z1_adj is the version adjusted so that 0 represents the average cutpoint for the dichotomous indicators composing the index; the _as_prob versions adjust these to the 0-1 range by applying stats::pnorm (treating them as probability measures).

se_z1

The standard error of the score.

_pct025, _pct975

The latent democracy score 95% confidence intervals.

num_measures

The number of measures that contributed to the index for the particular country-year. Varies from 1 to 31; one can filter the dataset to include only scores where more than, say, 5 measures have contributed to the index.

measures

The names of the specific measures that contributed to the index in that particular country-year. This is a list-column; use tidyr::unnest to unnest. See the documentation for generate_democracy_scores_dataset for specific info about any particular measure.

Standard descriptive variables (generated by this package)

extended_country_name

The name of the country in the Gleditsch-Ward system of states, or the official name of the entity (for non-sovereign entities and states not in the Gleditsch and Ward system of states) or else a common name for disputed cases that do not have an official name (e.g., Western Sahara, Hyderabad). The Gleditsch and Ward scheme sometimes indicates the common name of the country and (in parentheses) the name of an earlier incarnation of the state: thus, they have Germany (Prussia), Russia (Soviet Union), Madagascar (Malagasy), etc. For details, see Gleditsch, Kristian S. & Michael D. Ward. 1999. "Interstate System Membership: A Revised List of the Independent States since 1816." International Interactions 25: 393-413. The list can be found at http://privatewww.essex.ac.uk/~ksg/statelist.html.

GWn

Gleditsch and Ward's numeric country code, from the Gleditsch and Ward list of independent states.

cown

The Correlates of War numeric country code, 2016 version. This differs from Gleditsch and Ward's numeric country code in a few cases. See http://www.correlatesofwar.org/data-sets/state-system-membership for the full list.

in_GW_system

Whether the state is "in system" (that is, is independent and sovereign), according to Gleditsch and Ward, for this particular date. Matches at the end of the year; so, for example South Vietnam 1975 is FALSE because, according to Gleditsch and Ward, the country ended on April 1975 (being absorbed by North Vietnam). It is also TRUE for dates beyond 2012 for countries that did not end by then, depsite the fact that the Gleditsch and Ward list has not been updated since.

References

X. Marquez. "A Quick Method for Extending the Unified Democracy Scores". In: Available at SSRN 2753830 (2016). https://papers.ssrn.com/sol3/papers.cfm?abstract_id=2753830.

D. Pemstein, S. Meserve, and J. Melton. "Democratic Compromise: A Latent Variable Analysis of Ten Measures of Regime Type". In: Political Analysis 18.4 (2010), pp. 426-449. DOI: 10.1093/pan/mpq020.

See also

Author

Xavier Marquez, Political Science and International Relations Programme, Victoria University of Wellington, marquez.x@gmail.com