The Steffen Kailitz Dataset of Authoritarian Regime Types

Contains the authoritarian regime types dataset from S. Kailitz. "Classifying Political Regimes Revisited: Legitimation and Durability". In: Democratization 20.1 (2013), pp. 39-60.. This has been superceded by S. Kailitz. Varieties of Political Regimes (va-PoReg). Dataset. Dresden, 2024.. Use vaporeg for the most up to date version.

Usage

kailitz

Format

An object of class tbl_df (inherits from tbl, data.frame) with 9606 rows and 18 columns.

Source

S. Kailitz. "Classifying Political Regimes Revisited: Legitimation and Durability". In: Democratization 20.1 (2013), pp. 39-60.

Variables

year

The calendar year. Note that more than one regime type may obtain in a single year for a given country.

kailitz_country

The country name.

combined_regime

The "combined regime" for the country year. This is just the concatenation of all regimes for that year if the country has more than one regime type. For example, Spain has a "combined regime" of "Military Autocracy-One party Autocracy-Personalist Autocracy" for the each of the years 1946-1974.

kailitz_binary

Dichotomous democracy indicator. 1 if combined_regime is "Liberal democracy", 0 otherwise.

kailitz_tri

Trichotomous democracy indicator. 2 if combined_regime is "Liberal Democracy", 1 if it is "Electoral Autocracy" or "Electoral Autocracy-Liberal Democracy", 0 otherwise.

personal

Binary indicator of personalism. TRUE if combined_regime contains "Personal Autocracy", FALSE otherwise.

communist

Binary indicator of communism. TRUE if combined_regime contains "Communist Ideocracy", FALSE otherwise.

military

Binary indicator of military autocracy. TRUE if combined_regime contains "Military Autocracy", FALSE otherwise.

party

Binary indicator of one party autocracy TRUE if combined_regime contains "One party Autocracy", FALSE otherwise.

monarchy

Binary indicator of monarchy. TRUE if combined_regime contains "Monarchy", FALSE otherwise.

electoral

Binary indicator of electoral autocracy. TRUE if combined_regime contains "Electoral autocracy", FALSE otherwise.

failure

Binary indicator of state failure or occupation. TRUE if combined_regime contains "State Failure or Occupation", FALSE otherwise.

transition

Binary indicator of transition. TRUE if combined_regime contains "Transition", FALSE otherwise.

Standard descriptive variables (generated by this package)

extended_country_name

The name of the country in the Gleditsch-Ward system of states, or the official name of the entity (for non-sovereign entities and states not in the Gleditsch and Ward system of states) or else a common name for disputed cases that do not have an official name (e.g., Western Sahara, Hyderabad). The Gleditsch and Ward scheme sometimes indicates the common name of the country and (in parentheses) the name of an earlier incarnation of the state: thus, they have Germany (Prussia), Russia (Soviet Union), Madagascar (Malagasy), etc. For details, see Gleditsch, Kristian S. & Michael D. Ward. 1999. "Interstate System Membership: A Revised List of the Independent States since 1816." International Interactions 25: 393-413. The list can be found at http://privatewww.essex.ac.uk/~ksg/statelist.html.

GWn

Gleditsch and Ward's numeric country code, from the Gleditsch and Ward list of independent states.

cown

The Correlates of War numeric country code, 2016 version. This differs from Gleditsch and Ward's numeric country code in a few cases. See http://www.correlatesofwar.org/data-sets/state-system-membership for the full list.

in_GW_system

Whether the state is "in system" (that is, is independent and sovereign), according to Gleditsch and Ward, for this particular date. Matches at the end of the year; so, for example South Vietnam 1975 is FALSE because, according to Gleditsch and Ward, the country ended on April 1975 (being absorbed by North Vietnam). It is also TRUE for dates beyond 2012 for countries that did not end by then, depsite the fact that the Gleditsch and Ward list has not been updated since.

See also