Participation-Enhanced Polity Score

A dataset containing the Participation-Enhanced Polity score described in Bruce E. Moon, Jennifer Harvey Birdsall, Sylvia Ceisluk, Lauren M. Garlett, Joshua J. Hermias, Elizabeth Mendenhall, Patrick D. Schmid, and Wai Hong Wong (2006) "Voting Counts: Participation in the Measurement of Democracy" Studies in Comparative International Development 42, 2 (Summer, 2006). The complete dataset is available here: http://www.lehigh.edu/~bm05/democracy/Obtain_data.htm. The documentation below is based on the iformation in that website.

peps

Format

An object of class tbl_df (inherits from tbl, data.frame) with 15089 rows and 28 columns.

Source

Bruce E. Moon, Jennifer Harvey Birdsall, Sylvia Ceisluk, Lauren M. Garlett, Joshua J. Hermias, Elizabeth Mendenhall, Patrick D. Schmid, and Wai Hong Wong (2006) "Voting Counts: Participation in the Measurement of Democracy" Studies in Comparative International Development 42, 2 (Summer, 2006).

Details

peps_ccode

COW code, as used by the PEPS dataset.

year

Year.

polity1raw

Raw Polity1 score with missing data included, directly from Polity.

Polity1

Clean Polity1, irregular codes defined as missing data.

Polity2

Clean Polity2, irregular codes defined as missing data.

Polity3

Clean Preferred Polity score.

Pdemocracy

Clean Polity democracy, irregular codes defined as missing data.

Pautocracy

Clean Polity autocracy, irregular codes defined as missing data.

FHfree

Sum of Freedom House polrights & civliberties, perfect score is 12.

toI

IDEA Votes/Voting age population. n = 4966, missing = 11437.

toV

Vanhanen votes/two-thirds of Vanhanen population. n = 14441, missing = 1962.

toQ

(mostly)IDEA votes/Voting age population, with participation coded zero for noncompetitive elections. n = 8010, missing = 8393.

PEPS1i

PEPS1, using toI. n = 9163, missing = 7240.

PEPS2i

PEPS2, using toI. n = 4125, missing = 12278.

PEPS1q

PEPS1, using toQ. n = 9439, missing = 6964.

PEPS2q

PEPS2, using toQ. n = 6890, missing = 9513.

PEPS1v

PEPS1, using toV. n = 14090, missing = 2313.

PEPS2v

PEPS1, using toV. n = 13241, missing = 3162.

pop

Population in thousands, Penn World Tables.

rgdpch

Real per capita GDP, chain method, Penn World Tables.

suffrage

Legal suffrage as percent of Voting age population, from Paxton and Bollen.

cv

Compulsory voting. Coded as 4=strongly enforced,3=weakly enforced,2=NA,1=not enforced,0=none. Source: IDEA.

peps_country

The country names found in the original data. these had to be fixed - lots of them were blank.

Standard descriptive variables (generated by this package)

extended_country_name

The name of the country in the Gleditsch-Ward system of states, or the official name of the entity (for non-sovereign entities and states not in the Gleditsch and Ward system of states) or else a common name for disputed cases that do not have an official name (e.g., Western Sahara, Hyderabad). The Gleditsch and Ward scheme sometimes indicates the common name of the country and (in parentheses) the name of an earlier incarnation of the state: thus, they have Germany (Prussia), Russia (Soviet Union), Madagascar (Malagasy), etc. For details, see Gleditsch, Kristian S. & Michael D. Ward. 1999. "Interstate System Membership: A Revised List of the Independent States since 1816." International Interactions 25: 393-413. The list can be found at http://privatewww.essex.ac.uk/~ksg/statelist.html.

GWn

Gleditsch and Ward's numeric country code, from the Gleditsch and Ward list of independent states.

cown

The Correlates of War numeric country code, 2016 version. This differs from Gleditsch and Ward's numeric country code in a few cases. See http://www.correlatesofwar.org/data-sets/state-system-membership for the full list.

in_GW_system

Whether the state is "in system" (that is, is independent and sovereign), according to Gleditsch and Ward, for this particular date. Matches at the end of the year; so, for example South Vietnam 1975 is FALSE because, according to Gleditsch and Ward, the country ended on April 1975 (being absorbed by North Vietnam). It is also TRUE for dates beyond 2012 for countries that did not end by then, depsite the fact that the Gleditsch and Ward list has not been updated since.

See also