Download and process various democracy datasets.

Download and process various democracy datasets. Note that the datasets returned by the redownload_* family of functions (blm, bmr, bnr, gwf_autocratic, gwf_autocratic_extended, gwf_all, gwf_all_extended, LIED, magaloni, pacl, PIPE, peps, polyarchy, polyarchy_dimensions, uds_2014, uds_2010, uds_2011, ulfelder, utip, wahman_teorell_hadenius) are all available directly from this package and are unlikely to have changed since the package was installed. Access the respective dataset by typing its name, and refer to their documentation for details. You will not normally need to redownload them, unless you want to process the raw data yourself (set return_raw = TRUE) or suspect they have changed since the package was installed.

redownload_blm(url, verbose = TRUE, return_raw = FALSE, ...)

redownload_bmr(url, verbose = TRUE, return_raw = FALSE, ...)

redownload_bnr(url, verbose = TRUE, extend = FALSE, return_raw = FALSE,
  ...)

redownload_gwf(url, verbose = TRUE, extend = FALSE, dataset = c("all",
  "autocratic only"), return_raw = FALSE, ...)

redownload_lied(url, verbose = TRUE, return_raw = FALSE, ...)

redownload_pacl(url, verbose = TRUE, return_raw = FALSE, ...)

redownload_peps(url, verbose = TRUE, return_raw = FALSE, ...)

redownload_utip(url, verbose = TRUE, return_raw = FALSE, ...)

redownload_wahman_teorell_hadenius(url, verbose = TRUE, return_raw = FALSE,
  ...)

redownload_polyarchy_original(url, verbose = TRUE, return_raw = FALSE, ...)

redownload_polyarchy_dimensions(url, verbose = TRUE, return_raw = FALSE,
  ...)

redownload_magaloni(url, verbose = TRUE, extend = FALSE,
  return_raw = FALSE, ...)

redownload_uds(release_year = 2014, verbose = TRUE, return_raw = FALSE,
  ...)

redownload_ulfelder(url, verbose = TRUE, return_raw = FALSE,
  extend = FALSE, ...)

redownload_pipe(url, verbose = TRUE, return_raw = FALSE, ...)

Arguments

url
The URL of the dataset. This defaults to:

For blm: http://www.blmdemocracy.gatech.edu/blm final data.xls

For bmr: https://sites.google.com/site/mkmtwo/democracy-v2.0.dta?attredirects=0

For bnr: http://users.clas.ufl.edu/bernhard/content/data/meister1305.dta

For gwf_all and gwf_autocratic: http://sites.psu.edu/dictators/wp-content/uploads/sites/12570/2016/05/GWF-Autocratic-Regimes-1.2.zip

For LIED: https://dataverse.harvard.edu/api/access/datafile/2863101

For pacl: https://uofi.box.com/shared/static/bba3968d7c3397c024ec.dta

For peps: http://www.lehigh.edu/~bm05/democracy/PEPS1pub.dta

For utip: http://utip.lbj.utexas.edu/data/political regime data set RV.xls

For wahman_teorell_hadenius: https://sites.google.com/site/authoritarianregimedataset/data/ARD_V6.dta?attredirects=0&d=1

For polyarchy: https://www3.nd.edu/~mcoppedg/crd/poly8500.sav

For polyarchy_dimensions: http://www3.nd.edu/~mcoppedg/crd/DahlDims.sav

For magaloni: http://cddrl.fsi.stanford.edu/sites/default/files/res/Data_Set.xls

For UDS: for the 2014 release, it defaults to http://www.unified-democracy-scores.org/files/20140312/z/uds_summary.csv.gz; for the the 2011 release, to http://www.unified-democracy-scores.org/files/20110104/uds_summary.csv.gz; and for the 2010 release, to http://www.unified-democracy-scores.org/files/20100726/uds_summary.csv.gz

For ulfelder: https://dataverse.harvard.edu/api/access/datafile/2420018

For PIPE: https://sites.google.com/a/nyu.edu/adam-przeworski/home/data

verbose
Whether to print a running commentary of what the function is doing while processing the data.

return_raw
Whether to return the raw data, without any processing. Default is FALSE.

...
Other parameters passed to country_year_coder.

extend
(Only for redownload_bnr, redownload_gwf, redownload_magaloni, and redownload_ulfelder). Whether to extend the dataset back in time using a full panel of independent countries (for the redownload_bnr case) or the appropriate duration variable (gwf_duration, duration_nr, or rgjdura and rgjdurd, respectively, for redownload_gwf, redownload_magaloni, and redownload_ulfelder). For example, the United States enters the GWF dataset in 1946, where gwf_duration is already 75; one can extend the dataset to indicate that the country was classified as a democracy from 1872. Default is FALSE.

dataset
(Only for redownload_gwf). The dataset to output. Geddes, Wright, and Frantz provide two country-year files, one with autocratic regimes only ("autocratic only"), and one with both democratic and non-democratic regimes ("all"). Default is "all".

release_year
(Only in redownload_uds). The year of the UDS release to be downloaded. Can be 2014, 2011, or 2010. Default is 2014.

Source

Bowman, Kirk, Fabrice Lehoucq, and James Mahoney. 2005. Measuring Political Democracy: Case Expertise, Data Adequacy, and Central America. Comparative Political Studies 38 (8): 939-970. http://cps.sagepub.com/content/38/8/939. Data available at http://www.blmdemocracy.gatech.edu/.

Boix, Carles, Michael Miller, and Sebastian Rosato. 2012. A Complete Data Set of Political Regimes, 1800-2007. Comparative Political Studies 46 (12): 1523-1554. Available at https://sites.google.com/site/mkmtwo/data

Michael Bernhard, Timothy Nordstrom, and Christopher Reenock, "Economic Performance, Institutional Intermediation and Democratic Breakdown," Journal of Politics 63:3 (2001), pp. 775-803. Data and coding description available at http://users.clas.ufl.edu/bernhard/content/data/data.htm

Barbara Geddes, Joseph Wright, and Erica Frantz. 2014. "Autocratic Breakdown and Regime Transitions: A New Data Set." Perspectives on Politics 12(2): 313-331. The full data and codebook can be downloaded here http://sites.psu.edu/dictators/.

Skaaning, Svend-Erik; John Gerring; and Henrikas Bartusevicius (2015). "A Lexical Index of Electoral Democracy." Comparative Political Studies, Vol. 48, No. 12, pp. 1491-1525. Original data and variable descriptions available at https://dataverse.harvard.edu/dataset.xhtml?persistentId=doi:10.7910/DVN/29106

Cheibub, Jose Antonio, Jennifer Gandhi, and James Raymond Vreeland. 2010. "Democracy and Dictatorship Revisited." Public Choice, vol. 143, no. 2-1, pp. 67-101. DOI: 10.1007/s11127-009-9491-2. The full data and codebook can be downloaded here https://sites.google.com/site/joseantoniocheibub/datasets/democracy-and-dictatorship-revisited

Bruce E. Moon, Jennifer Harvey Birdsall, Sylvia Ceisluk, Lauren M. Garlett, Joshua J. Hermias, Elizabeth Mendenhall, Patrick D. Schmid, and Wai Hong Wong (2006) "Voting Counts: Participation in the Measurement of Democracy" Studies in Comparative International Development 42, 2 (Summer, 2006). The complete dataset is available here: http://www.lehigh.edu/~bm05/democracy/Obtain_data.htm.

The University of Texas Inequality Project Categorical Dataset of Political Regimes. Described in Sara Hsu, "The Effect of Political Regimes on Inequality, 1963-2002," UTIP Working Paper No. 53 (2008), http://utip.gov.utexas.edu/papers/utip_53.pdf. Data available for download at http://utip.gov.utexas.edu/data/

Wahman, Michael, Jan Teorell, and Axel Hadenius. 2013. Authoritarian regime types revisited: updated data in comparative perspective. Contemporary Politics 19 (1): 19-34. The dataset and codebook can be downloaded from https://sites.google.com/site/authoritarianregimedataset/data

Michael Coppedge and Wolfgang Reinicke, "Measuring Polyarchy," Studies in Comparative International Development 25:1 (Spring 1990): 51-72. Data available at http://www3.nd.edu/~mcoppedg/crd/datacrd.htm

Michael Coppedge, Angel Alvarez, and Claudia Maldonado, "Two Persistent Dimensions of Democracy: Contestation and Inclusiveness," Journal of Politics 70:3 (July 2008): 632-647.

Magaloni, Beatriz, Jonathan Chu, and Eric Min. 2013. Autocracies of the World, 1950-2012 (Version 1.0). Dataset, Stanford University. Original data and codebook available at http://cddrl.fsi.stanford.edu/research/autocracies_of_the_world_dataset/.

Pemstein, Daniel, Stephen Meserve, and James Melton. 2010. Democratic Compromise: A Latent Variable Analysis of Ten Measures of Regime Type. Political Analysis 18 (4): 426-449. Original data and codebook available at http://www.unified-democracy-scores.org/.

Jay Ulfelder. 2012. Democracy/Autocracy Data Set. http://hdl.handle.net/1902.1/18836.

Adam Przeworski. 2013 Political Institutions and Political Events (PIPE) Data Set. Data set. https://sites.google.com/a/nyu.edu/adam-przeworski/home/data.

Value

A tibble with the processed dataset, unless return_raw is TRUE, in which case the function returns the raw data without processing.

Examples

# NOT RUN {
redownload_blm()
# }# NOT RUN {
bmr <- redownload_bmr()
bmr
# }# NOT RUN {
redownload_bnr()
redownload_bnr(full_panel = FALSE)
# }# NOT RUN {
gwf <- redownload_gwf()
gwf
# }# NOT RUN {
redownload_lied()
# }# NOT RUN {
redownload_pacl()
# }# NOT RUN {
redownload_peps()
# }# NOT RUN {
redownload_utip()
# }# NOT RUN {
redownload_wahman_teorell_hadenius()
# }# NOT RUN {
redownload_polyarchy_original()
# }# NOT RUN {
redownload_polyarchy_dimensions()
# }# NOT RUN {
redownload_magaloni()
# }# NOT RUN {
redownload_uds()
redownload_uds(2011)
# }# NOT RUN {
redownload_ulfelder()
# }# NOT RUN {
redownload_pipe()
# }