Download and process various democracy datasets.

Download and process various democracy datasets. Note that the datasets returned by the redownload_* family of functions (blm, bmr, bnr, bti, gwf_autocratic, gwf_autocratic_extended, gwf_all, gwf_all_extended, LIED, magaloni, pacl, pacl_update, PIPE, peps, polityIV, polyarchy, polyarchy_dimensions, REIGN, uds_2014, uds_2010, uds_2011, ulfelder, utip, wahman_teorell_hadenius, anckar, svmdi) are all available directly from this package and are unlikely to have changed since the package was installed. Access the respective dataset by typing its name, and refer to their documentation for details. You will not normally need to redownload them, unless you want to process the raw data yourself (set return_raw = TRUE) or suspect they have changed since the package was installed.

redownload_blm(url, verbose = TRUE, return_raw = FALSE, ...)

redownload_anckar(url, verbose = TRUE, return_raw = FALSE, ...)

redownload_bmr(url, verbose = TRUE, return_raw = FALSE, ...)

redownload_gwf(
  url,
  verbose = TRUE,
  extend = FALSE,
  dataset = c("all", "autocratic only"),
  return_raw = FALSE,
  ...
)

redownload_pacl(url, verbose = TRUE, return_raw = FALSE, ...)

redownload_peps(url, verbose = TRUE, return_raw = FALSE, ...)

redownload_utip(url, verbose = TRUE, return_raw = FALSE, ...)

redownload_wahman_teorell_hadenius(
  url,
  verbose = TRUE,
  return_raw = FALSE,
  ...
)

redownload_polyarchy_original(url, verbose = TRUE, return_raw = FALSE, ...)

redownload_polyarchy_dimensions(url, verbose = TRUE, return_raw = FALSE, ...)

redownload_magaloni(
  url,
  verbose = TRUE,
  extend = FALSE,
  return_raw = FALSE,
  ...
)

redownload_svmdi(
  url,
  release_year = 2020,
  verbose = TRUE,
  return_raw = FALSE,
  ...
)

redownload_ulfelder(
  url,
  verbose = TRUE,
  return_raw = FALSE,
  extend = FALSE,
  ...
)

redownload_pipe(url, verbose = TRUE, return_raw = FALSE, ...)

redownload_bti(url, verbose = TRUE, return_raw = FALSE, ...)

redownload_pacl_update(url, verbose = TRUE, return_raw = FALSE, ...)

redownload_reign(url, verbose = TRUE, return_raw = FALSE, ...)

Source

K. Bowman, F. Lehoucq, and J. Mahoney. "Measuring Political Democracy: Case Expertise, Data Adequacy, and Central America". In: Comparative Political Studies 38.8 (2005), pp. 939-970. DOI: 10.1177/0010414005277083. Data used to be available at http://www.blmdemocracy.gatech.edu/.

C. Anckar and C. Fredriksson. "Classifying political regimes 1800-2016: a typology and a new dataset". In: European Political Science (2018). DOI: 10.1057/s41304-018-0149-8. https://doi.org/10.1057/s41304-018-0149-8. Updated data (V2) available at https://dataverse.harvard.edu/file.xhtml?persistentId=doi:10.7910/DVN/7SSSAH/DIZXSI&version=2.0.

C. Boix, M. Miller, and S. Rosato. "A Complete Dataset of Political Regimes, 1800-2007". In: Comparative Political Studies 46.12 (2012), pp. 1523-1554. DOI: 10.1177/0010414012463905. Available at https://sites.google.com/site/mkmtwo/data

B. Geddes, J. Wright, and E. Frantz. "Autocratic Breakdown and Regime Transitions: A New Data Set". In: Perspectives on Politics 12.1 (2014), pp. 313-331. DOI: 10.1017/S1537592714000851. The full data and codebook can be downloaded here: http://sites.psu.edu/dictators/.

J. Cheibub, J. Gandhi, and J. Vreeland. "Democracy and dictatorship revisited". In: Public Choice 143.1 (2010), pp. 67-101. DOI: 10.1007/s11127-009-9491-2. The full data and codebook can be downloaded here https://sites.google.com/site/joseantoniocheibub/datasets/democracy-and-dictatorship-revisited

B. E. Moon, J. H. Birdsall, S. Ciesluk, et al. "Voting Counts: Participation in the Measurement of Democracy". In: Studies in Comparative International Development 41.2 (2006), pp. 3-32. DOI: 10.1007/BF02686309. The complete dataset is available here: http://www.lehigh.edu/~bm05/democracy/Obtain_data.htm.

S. Hsu. "The Effect of Political Regimes on Inequality, 1963-2002". In: UTIP Working Paper (2008). Data available for download at http://utip.gov.utexas.edu/data/

M. Wahman, J. Teorell, and A. Hadenius. "Authoritarian Regime Types Revisited: Updated Data in Comparative Perspective". In: Contemporary Politics 19.1 (2013), pp. 19-34. https://sites.google.com/site/authoritarianregimedataset/data. The dataset and codebook can be downloaded from https://sites.google.com/site/authoritarianregimedataset/data

M. Coppedge and W. H. Reinicke. "Measuring Polyarchy". In: Studies in Comparative International Development 25.1 (1990), pp. 51-72. DOI: 10.1007/Bf02716905. Data available at http://www3.nd.edu/~mcoppedg/crd/datacrd.htm

M. Coppedge, A. Alvarez, and C. Maldonado. "Two Persistent Dimensions of Democracy: Contestation and Inclusiveness". In: The journal of politics 70.03 (2008), pp. 632-647. DOI: 10.1017/S0022381608080663.

B. Magaloni, J. Chu, and E. Min. Autocracies of the World, 1950-2012 (Version 1.0). Dataset. 2013. http://cddrl.fsi.stanford.edu/research/autocracies_of_the_world_dataset. Original data and codebook used to be available at the link.

K. Gründler and T. Krieger. Machine Learning Indices, Political Institutions, and Economic Development. Report. CESifo Group Munich, 2018. https://www.cesifo-group.de/DocDL/cesifo1_wp6930.pdf. Original data available at https://ml-democracy-index.net/.

J. Ulfelder. Democracy/Autocracy Data Set. 2012. http://hdl.handle.net/1902.1/18836.

A. Przeworski. Political Institutions and Political Events (PIPE) Data Set. Data set. 2013. https://sites.google.com/a/nyu.edu/adam-przeworski/home/data. https://sites.google.com/a/nyu.edu/adam-przeworski/home/data.

Bertelsmann Stiftung. Transformation Index of the Bertelsmann Stiftung 2022. Tech. rep. Bertelsmann Stiftung, 2022. Available at https://bti-project.org/en/index/political-transformation

C. Bjørnskov and M. Rode. "Regime types and regime change: A new dataset on democracy, coups, and political institutions". In: The Review of International Organizations 15.2 (2020), pp. 531-551. DOI: 10.1007/s11558-019-09345-1. Available at http://www.christianbjoernskov.com/bjoernskovrodedata/

C. Bell. The Rulers, Elections, and Irregular Governance Dataset (REIGN). 2016. http://oefresearch.org/datasets/reign. Available at http://oefresearch.org/datasets/reign

Arguments

url

The URL of the dataset. This defaults to:

For anckar: https://dataverse.harvard.edu/api/access/datafile/3868064?format=original&gbrecs=true

For blm: http://www.blmdemocracy.gatech.edu/blm%20final%20data.xls . This doesn't work anymore.

For bmr: https://dataverse.harvard.edu/api/access/datafile/3130643

For bnr: http://users.clas.ufl.edu/bernhard/content/data/meister1305.dta . This doesn't work anymore.

For bti: https://bti-project.org/fileadmin/api/content/en/downloads/data/BTI_2006-2022_Scores.xlsx

For gwf_all and gwf_autocratic: http://sites.psu.edu/dictators/wp-content/uploads/sites/12570/2016/05/GWF-Autocratic-Regimes-1.2.zip

For LIED: https://dataverse.harvard.edu/api/access/datafile/:persistentId?persistentId=doi:10.7910/DVN/29106/SXRLK1

For pacl: https://uofi.box.com/shared/static/bba3968d7c3397c024ec.dta

For pacl_update http://www.christianbjoernskov.com/wp-content/uploads/2020/09/Bj%C3%B8rnskov-Rode-integrated-dataset-v3.2.xlsx

For peps: http://www.lehigh.edu/~bm05/democracy/PEPS1pub.dta

For svmdi: https://www.dropbox.com/s/a7yqs5txt3qpwn0/Index%20Upload.xlsx?dl=1.

For the 2016 release, it defaults to http://www.wiwi.uni-wuerzburg.de/fileadmin/12010400/Data.dta

For REIGN: https://github.com/OEFDataScience/REIGN.github.io/blob/gh-pages/data_sets/regime_list.csv?raw=true

For utip: http://utip.lbj.utexas.edu/data/political%20regime%20data%20set%20RV.xls

For wahman_teorell_hadenius: https://sites.google.com/site/authoritarianregimedataset/data/ARD_V6.dta?attredirects=0&d=1

For polyarchy: https://www3.nd.edu/~mcoppedg/crd/poly8500.sav

For polyarchy_dimensions: http://www3.nd.edu/~mcoppedg/crd/DahlDims.sav

For magaloni: https://fsi-live.s3.us-west-1.amazonaws.com/s3fs-public/res/Data_Set.xls

For ulfelder: https://dataverse.harvard.edu/api/access/datafile/2420018

For PIPE: https://sites.google.com/a/nyu.edu/adam-przeworski/home/data . This doesn't work anymore.

verbose

Whether to print a running commentary of what the function is doing while processing the data.

return_raw

Whether to return the raw data, without any processing. Default is FALSE.

...

Other parameters passed to country_year_coder.

extend

(Only for redownload_gwf, redownload_magaloni, and redownload_ulfelder). Whether to extend the dataset back in time using the appropriate duration variable (gwf_duration, duration_nr, or rgjdura and rgjdurd, respectively, for redownload_gwf, redownload_magaloni, and redownload_ulfelder). For example, the United States enters the GWF dataset in 1946, where gwf_duration is already 75; one can extend the dataset to indicate that the country was classified as a democracy from 1872. Default is FALSE.

dataset

(Only for redownload_gwf). The dataset to output. Geddes, Wright, and Frantz provide two country-year files, one with autocratic regimes only ("autocratic only"), and one with both democratic and non-democratic regimes ("all"). Default is "all".

release_year

(Only in redownload_svmdi). The year of the release to be downloaded. For svmdi, it can be 2016 or 2020.

Value

A tibble with the processed dataset, unless return_raw

is TRUE, in which case the function returns the raw data without processing.

Examples